### 2023-06-06 #212 @whedon
[PRE REVIEW]: An R Companion for Introduction to Data Mining
**Submitting author:** @mhahsler (<a href="http://orcid.org/0000-0003-2716-1405">Michael Hahsler</a>)
**Repository:** https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
