### 2023-06-19 2ddcdea
Remove AAS info
