### 2023-06-20 #212 @mhahsler
@xuanxu the pdf generation script seems to die when adding an image (not finding the journal logo?)

```
Error running filter /usr/local/share/openjournals/data/filters/add-images.lua:
File  not found in resource path
stack traceback:
	/usr/local/share/openjournals/data/filters/add-images.lua:4: in upvalue 'insert'
	/usr/local/share/openjournals/data/filters/add-images.lua:9: in function 'Meta'
```

I am not sure how to fix this.
