### 2023-06-20 #212 @xuanxu
Yes, is something we are looking into, we've updated the PDF compilation step and this is probably some misconfiguration in our part.
