### 2023-08-20 #212 @mhahsler
@labarba Hi. There was a technical issue with creating the pdf. The issue was resolved. What are the next steps for the review process? Best regards, Michael
