+++ book_src/render.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env Rscript
+
+bookdown::render_book("index.Rmd", "bookdown::bs4_book")
