+++ book_src/packages.bib
@@ -1,9 +1,9 @@
 @Manual{R-arules,
   title = {arules: Mining Association Rules and Frequent Itemsets},
   author = {Michael Hahsler and Christian Buchta and Bettina Gruen and Kurt Hornik},
-  year = {2023},
-  note = {R package version 1.7-6},
+  note = {R package version 1.7-6.1},
   url = {https://github.com/mhahsler/arules},
+  year = {2023},
 }
 
 @Manual{R-arulesViz,
@@ -49,12 +49,12 @@ al.},
 }
 
 @Manual{R-dbscan,
-  title = {dbscan: Density-Based Spatial Clustering of Applications with Noise
-(DBSCAN) and Related Algorithms},
+  title = {dbscan: Density-Based Spatial Clustering of Applications with Noise (DBSCAN)
+and Related Algorithms},
   author = {Michael Hahsler and Matthew Piekenbrock},
-  year = {2022},
-  note = {R package version 1.1-11},
+  note = {R package version 1.1-11-1},
   url = {https://github.com/mhahsler/dbscan},
+  year = {2023},
 }
 
 @Manual{R-dplyr,
@@ -166,16 +166,16 @@ Theory Group (Formerly: E1071), TU Wien},
   title = {MASS: Support Functions and Datasets for Venables and Ripley's MASS},
   author = {Brian Ripley},
   year = {2023},
-  note = {R package version 7.3-59},
+  note = {R package version 7.3-60},
   url = {http://www.stats.ox.ac.uk/pub/MASS4/},
 }
 
 @Manual{R-Matrix,
   title = {Matrix: Sparse and Dense Matrix Classes and Methods},
   author = {Douglas Bates and Martin Maechler and Mikael Jagan},
-  year = {2022},
-  note = {R package version 1.5-1},
-  url = {https://CRAN.R-project.org/package=Matrix},
+  year = {2023},
+  note = {R package version 1.6-0/r4087},
+  url = {https://R-Forge.R-project.org/projects/matrix/},
 }
 
 @Manual{R-mclust,
@@ -198,16 +198,16 @@ Classification, and Density Estimation},
 @Manual{R-nnet,
   title = {nnet: Feed-Forward Neural Networks and Multinomial Log-Linear Models},
   author = {Brian Ripley},
-  year = {2022},
-  note = {R package version 7.3-18},
+  year = {2023},
+  note = {R package version 7.3-19},
   url = {http://www.stats.ox.ac.uk/pub/MASS4/},
 }
 
 @Manual{R-plotly,
   title = {plotly: Create Interactive Web Graphics via plotly.js},
   author = {Carson Sievert and Chris Parmer and Toby Hocking and Scott Chamberlain and Karthik Ram and Marianne Corvellec and Pedro Despouy},
-  year = {2022},
-  note = {R package version 4.10.1},
+  year = {2023},
+  note = {R package version 4.10.2},
   url = {https://CRAN.R-project.org/package=plotly},
 }
 
@@ -215,7 +215,7 @@ Classification, and Density Estimation},
   title = {pROC: Display and Analyze ROC Curves},
   author = {Xavier Robin and Natacha Turck and Alexandre Hainard and Natalia Tiberti and Frédérique Lisacek and Jean-Charles Sanchez and Markus Müller},
   year = {2023},
-  note = {R package version 1.18.2},
+  note = {R package version 1.18.4},
   url = {http://expasy.org/tools/pROC/},
 }
 
@@ -296,16 +296,16 @@ Regression},
   title = {scatterpie: Scatter Pie Plot},
   author = {Guangchuang Yu},
   year = {2023},
-  note = {R package version 0.2.0},
+  note = {R package version 0.2.1},
[…]
