### 2024-01-08 #212 @mhahsler
@stats-tgeorge Thank you for your reply. It has been a while, but we are still teaching the course and I am still interested having the companion reviewed.
