### 2024-01-08 mhahsler/Introduction_to_Data_Mining_R_Examples#5 @stats-tgeorge
JOSE Submission JOSE Paper
JOSE Paper Updates
- [x] Describe the submission, and explain its eligibility for JOSE.
- [x] Tell us the “story” of the project: how did it come to be?
- [x] Cite key references, including a link to the open archive of the software or the learning module. You likely need to call your .bib file.
