### 2024-01-08 #212 @stats-tgeorge
@mhahsler great! To start, looking at your JOSE paper there are some missing pieces. See https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples/issues/5#issue-2071103726
