### 2024-01-22 #212 @stats-tgeorge
Hello @Hughshanahan, Michael Hahsler has submitted a data mining companion book to JOSE for other instructors to use. Would you be willing to review the submission? Your time and efforts will be greatly appreciated!
