### 2024-02-13 #223 @editorialbot
[REVIEW]: An R Companion for Introduction to Data Mining
**Submitting author:** @mhahsler (<a href="http://orcid.org/0000-0003-2716-1405">Michael Hahsler</a>)
**Repository:** https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
**Archive:** 10.6084/m9.figshare.26750404.v3
