### 2024-05-22 #223 @rudeboybert
So sorry @stats-tgeorge to have dropped the ball on this. It was a brutal semester. Do you still need this? If so, I can get it to you by the evening of Thu 6/30.
