### 2024-05-24 #212 @stats-tgeorge
Hello @trallard, you are on our list of potential reviewers for JOSE. Michael Hahsler has submitted a data mining companion book to JOSE for other instructors to use. Would you be willing to review the submission? Your time and efforts will be greatly appreciated!
