### 2024-06-17 #223 @mhahsler
@rudeboybert Thank you for the review. I have addressed the three issues:

1. Added a section called "Software Requirements" in README.md with installation instructions.
2. Added a section called "Statement of Need" in README.md
3. Created a release with version number 1.0.0 for the GitHub repository.

I am not sure if creating the release is sufficient to address "Version: Does the release version given match the repository release?".

Best regards,
Michael
