+++ README.md
@@ -22,11 +22,11 @@ for corrections or to suggest improvements.
 | Chapter | Slides | R Code Companion | Sample Textbook Chapters|  
 | :--------| :---: | :----: | :--: |
 | 1. Introduction | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap1_intro.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/introduction.html) | |
-| 2. Data | Slides: [Data](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pdf), [Exploration](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_exploration.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html) | |
-| 3. Classification: Basic Concepts and Techniques | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap3_basic_classification.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-basic-concepts-and-techniques.html) | [Read Chapter 3](https://www-users.cs.umn.edu/~kumar001/dmbook/ch3_classification.pdf) |
+| 2. Data | Slides: [Data](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pdf), [Exploration](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_exploration.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html) | [Data Exploration (Web Chapter)](https://www-users.cse.umn.edu/~kumar001/dmbook/data_exploration_1st_edition.pdf) |
+| 3. Classification: Basic Concepts and Techniques | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap3_basic_classification.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-basic-concepts-and-techniques.html) | [Chapter 3](https://www-users.cs.umn.edu/~kumar001/dmbook/ch3_classification.pdf) |
 | 4. Classification: Alternative Techniques | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap4_alternative_classification.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-alternative-techniques.html) | |
-| 5. Association Analysis: Basic Concepts and Algorithms | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap5_basic_association_analysis.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/association-analysis-basic-concepts-and-algorithms.html) | [Read Chapter 5](https://www-users.cs.umn.edu/~kumar001/dmbook/ch5_association_analysis.pdf) |
[…]
