### 2024-06-21 mhahsler/Introduction_to_Data_Mining_R_Examples#5 @stats-tgeorge
@mhahsler Can you confirm these were completed a while back?
