### 2024-06-21 mhahsler/Introduction_to_Data_Mining_R_Examples#5 @mhahsler
@stats-tgeorge Yes, the paper contains the eligibility and the story in the paper section "Statement of Need." There is a link to the GitHub version of the book at the end of the summary. Please let me know if additional changes are needed.
