### 2024-06-21 #223 @mhahsler
The proof looks good.

@stats-tgeorge  Thank you for guiding me through this process. It has helped me with improving the resource.
