### 2024-06-21 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @stats-tgeorge
Review Items
@mhahsler

First of all, there is so much content here. Thank you for your work.

- [ ] Statement of need: Please specify the target audience. Consider different levels of courses the materials could be considered for. A first course in statistics or data science? A senior undergraduate course? A graduate course? Suggested pre-requisites?
- [ ] Usage: In your JOSE paper, discuss how instructions can use all or parts of this for teaching. You discuss how self-learners might use it but not how someone could adopt all or parts of the materials into their class.

**Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)**
- [ ] Learning objectives: You give a brief outline in Chapter 1 of the topics to be covered in your companion book. I would like to see more depth. Is this going to be application-focused? Or, will both the theory and application be presented? Consider what would be useful to see to determine if someone should keep reading. Many books have these topics, what will they find about them, and how to implement them with your companion?
- [ ] Content quality:
  - Chapter 4 is missing any description of the outputs of the classification models. Specifically 4.3.4, 4.3.5, 4.3.6, 4.3.7, 4.3.8,
  - The same is true in 4.7, with no comments. There are also many conditions that need to be met in order to use many of these approaches.
  - In 5.5.3 there is a coding warning but it is not discussed and appears to be important.
  - In 5.7 the interactive table is overlapping with your table of contents on the right.

**JOSE Paper**

- [ ] Statment of need: Same as above.
- [ ] Description: Please add more comments on the organization. You have the online book and webpage with a table of materials people can click on as wanted. Give an outline of the materials available, and what each resource is for.
- [ ] References: DOI - see the mentioned failed DOIs in the editorial bot response on the review.
- [ ] Story - add a bit of the background story from your book into the JOSE paper

General Comments (do not have to be addressed)
- Consider an easier introduction to the dplyr/tidyverse verbs. Jumping directly to using something like across() within a summarize before showing examples of one or the other will be challenging for readers.
- Consider including _Palmer Penguins_ to diversity your examples on unsupervised techniques (or supervised with grouping).
- I would rename your 2.11.2 section, grouped operations or calculations.
- As nice as it is to put your packages all in 1 place, you might consider putting packages with very specific uses in those specific sections. This will allow easier copy/paste/use.
- I'm assuming your ordering is due to the book ordering. If not, the order should be 2.11 -> 2.12 both before 2.7 or 2.8.
- In future versions, _tidymodels_ has come a long way and can help to unify much of the code under more similar syntax. The author of Caret, Max Kuhn, now is also working on _tidymodels_.  I see you gave that as a future reference.

-
