### 2024-06-21 #223 @stats-tgeorge
@mhahsler I've added a to-do list as an issue. [Here](https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples/issues/6)
