+++ paper/paper.bib
@@ -20,7 +20,7 @@
   YEAR         = 2005,
   VOLUME       = 14,
   NUMBER       = 15,
-  url          = {http://www.jstatsoft.org/v14/i15/},
+  doi          = {10.18637/jss.v014.i15},
   MONTH        = {October},
   PAGES        = {1--25},
   ISSN         = {1548-7660},
@@ -86,4 +86,13 @@
   author = {{Kuhn} and {Max}},
   year = {2008},
   pages = {1–26},
+}
+
+@Manual{tidymodels2020,
+    title = {Tidymodels: a collection of packages for modeling and
+      machine learning using tidyverse principles.},
+    author = {Max Kuhn and Hadley Wickham},
+    url = {https://www.tidymodels.org},
+    year = {2020},
+    doi = {10.32614/CRAN.package.tidymodels}
 }
