+++ paper/paper.md
@@ -1,74 +1,124 @@
 ---
 title: "An R Companion for Introduction to Data Mining"
 tags:
-  - R
-  - data mining
+- R
+- data mining
+date: "30 May 2023"
+output: pdf_document
 authors:
-  - name: Michael Hahsler
-    orcid: 0000-0003-2716-1405
-    affiliation: 1
-affiliations:
- - name: Southern Methodist University
-   index: 1
-date: 30 May 2023
+- name: Michael Hahsler
+  orcid: "0000-0003-2716-1405"
+  affiliation: 1
 bibliography: paper.bib
+affiliations:
+- name: Southern Methodist University
+  index: 1
 ---
 
 # Summary
 
 An R Companion for Introduction to Data Mining is an open-source learning and 
-teaching resource for an introductory course in data mining. It
-can be used to accompany the popular
+teaching resource that covers how to implement data mining concepts using R.
+It can be used to accompany the popular
 data mining textbook Introduction to Data Mining [@Tan2018] or as a stand-alone resource
-to study the basic data mining concepts including data preparation, classification,
-clustering, and association analysis. The resource uses complete annotated examples 
-to teach how the basic concepts are translated into R code.
+to study the implementation of the basic data mining concepts 
+including data preparation, classification,
+clustering, and association analysis. 
+The resource uses complete annotated examples 
+to demonstrate how data mining concepts are translated into R code.
+
 
 The materials have been made publicly available at: <https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/> and licensed under the [Creative Commons Attribution-NonCommercial-ShareAlike 4.0 (CC BY-NC-SA 4.0) License](https://creativecommons.org/licenses/by-nc-sa/4.0/).
 
 # Statement of Need
 
 The textbook Introduction to Data Mining has been one of the most popular choices
-to learn data mining concepts. Several of the key chapters have been made available for free by the authors [@Tan2018].
+to learn data mining concepts.
+Several of the key chapters have been made available for free by the authors [@Tan2018].
 One of the authors also provides Python Jupyter notebooks with examples, but
 complete R code examples were missing. Given the R community's interest in 
-data analysis, data science, and machine learning, and the wide support of R packages for data mining, 
-there was clearly a gap that needed to be filled.
+data analysis, data science, and machine learning, and the wide support of R packages for data mining, there was an obvious gap that is filled by this learning resource.
+This resource targets advanced undergraduate and graduate student and can be used as a first
+introduction to data mining. 
 
-During many years of teaching data mining with R, I developed the Companion for Introduction to Data Mining
-resource mainly based on the popular tidyverse package collection [@tidyverse2019], caret [@caret2008], and a set of packages
-developed with students to better support different data mining tasks (arules [@arules2005], seriation [@seriation2008]
-arulesViz [@arulesViz2017], and dbscan [@dbscan2019]).
+# Learning Objectives and Content
+
+The resource assumes basic knowledge of programming and statistics.
+The learning objectives are to
+learn how to
+
+* prepare and understand data,
+* perform classification,
+* perform association analysis, and
+* perform cluster analysis
+
+using self-contained and annotated R code examples that work with small datasets
+carefully chosen to show the learner important aspects of data mining. 
+The learner can copy-and-paste the examples into an R markdown notebook to experiment with 
+the code and the provided example data and then modify the code to work on their own data sets.
+This learning-by-doing approach has worked very well in preparing students to work
+with more complex real-world datasets by relieving them from dealing with 
+too many implementation details while exploring the concepts. 
+
+To make using the resource easier, the resources mirrors 
+the structure of the textbook. After a short introduction, 
[…]
