+++ book_src/07_cluster_analysis.Rmd
@@ -4,24 +4,26 @@ editor_options:
     wrap: 72
 ---
 
-# Clustering Analysis
+# Cluster Analysis
 
 ```{r setup_07}
-pkgs <- sort(c('tidyverse', 'factoextra', 'dbscan', 'cluster', 'mclust', 
-  'kernlab', 'e1071', 'scatterpie', 'fpc', 'seriation', 'mlbench', 'GGally'
-))
+pkgs <- c("cluster", "dbscan", "e1071", "factoextra", "fpc", 
+          "GGally", "kernlab", "mclust", "mlbench", "scatterpie", 
+          "seriation", "tidyverse")
   
 pkgs_install <- pkgs[!(pkgs %in% installed.packages()[,"Package"])]
 if(length(pkgs_install)) install.packages(pkgs_install)
 ```
 
-The packages used for this chapter are: `r format_pkgs(pkgs)`
-
 ```{r setup_07-2, include=FALSE}
+source("_common.R")
+source("format_pkgs.R")
 all_pkgs <- union(all_pkgs, pkgs)
 ```
 
-## Introduction
+The packages used for this chapter are: `r format_pkgs(pkgs)`
+
+## Overview
 
 [Cluster analysis or clustering](https://en.wikipedia.org/wiki/Cluster_analysis) 
 is the task of grouping a set of objects in such a way that objects in the same group (called a cluster) are more similar (in some sense) to each other than to those in other groups (clusters).
@@ -34,7 +36,7 @@ You can read the free sample chapter from the textbook [@Tan2005]:
 [Chapter 7. Cluster Analysis: Basic Concepts and
 Algorithms](https://www-users.cs.umn.edu/~kumar001/dmbook/ch7_clustering.pdf)
 
-## Data Preparation
+### Data Preparation
 
 ```{r }
 library(tidyverse)
@@ -80,14 +82,19 @@ that this is not an issue for the Ruspini dataset with both, x and y,
 being roughly between 0 and 150. Most data analysts will still scale
 each column in the data to zero mean and unit standard deviation
 ([z-scores](https://en.wikipedia.org/wiki/Standard_score)). 
+
 *Note:* The standard `scale()` function scales a whole data
 matrix so we implement a function for a single vector and apply it to
 all numeric columns.
 
 ```{r }
 ## I use this till tidyverse implements a scale function
-scale_numeric <- function(x) mutate_if(x, is.numeric, function(y) as.vector(scale(y)))
+scale_numeric <- function(x) {
+  x |> mutate(across(where(is.numeric), function(y) as.vector(scale(y))))
+}
+```
 
+```{r }
 ruspini_scaled <- ruspini |> 
   scale_numeric()
 summary(ruspini_scaled)
@@ -97,9 +104,8 @@ After scaling, most z-scores will fall in the range $[-3,3]$ (z-scores
 are measured in standard deviations from the mean), where $0$ means
 average.
 
-## Clustering methods
 
-### k-means Clustering
+## K-means
 
 [k-means](https://en.wikipedia.org/wiki/K-means_clustering) implicitly
 assumes Euclidean distances. We use $k = 4$ clusters and run the
@@ -117,27 +123,34 @@ using `km$cluster`. I add the cluster assignment as a column to the
 scaled dataset (I make it a factor since it represents a nominal label).
 
 ```{r }
-ruspini_clustered <- ruspini_scaled |> add_column(cluster = factor(km$cluster))
+ruspini_clustered <- ruspini_scaled |> 
+  add_column(cluster = factor(km$cluster))
 ruspini_clustered
 
[…]
