+++ book_src/06_arules.Rmd
@@ -1,3 +1,4 @@
 # Association Analysis: Advanced Concepts
 
-No code is available for this Chapter. Some topics were already covered in the code for the previous chapter.
+No code is available for this Chapter. Some topics were already covered in the 
+code for the previous chapter.

[…]
