+++ paper/paper.md
@@ -19,13 +19,13 @@ affiliations:
 
 An R Companion for Introduction to Data Mining is an open-source learning and 
 teaching resource that covers how to implement data mining concepts using R.
-It can be used to accompany the popular
-data mining textbook Introduction to Data Mining [@Tan2018] or as a stand-alone resource
+It is designed to accompany the popular
+data mining textbook _Introduction to Data Mining_ [@Tan2018]
 to study the implementation of the basic data mining concepts 
 including data preparation, classification,
 clustering, and association analysis. 
-The resource uses complete annotated examples 
-to demonstrate how data mining concepts are translated into R code.
+The resource uses complete, annotated examples 
+to demonstrate how data mining concepts can be translated into R code.
 
 
 The materials have been made publicly available at: <https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/> and licensed under the [Creative Commons Attribution-NonCommercial-ShareAlike 4.0 (CC BY-NC-SA 4.0) License](https://creativecommons.org/licenses/by-nc-sa/4.0/).
@@ -36,9 +36,10 @@ The textbook Introduction to Data Mining has been one of the most popular choice
 to learn data mining concepts.
 Several of the key chapters have been made available for free by the authors [@Tan2018].
 One of the authors also provides Python Jupyter notebooks with examples, but
-complete R code examples were missing. Given the R community's interest in 
-data analysis, data science, and machine learning, and the wide support of R packages for data mining, there was an obvious gap that is filled by this learning resource.
-This resource targets advanced undergraduate and graduate student and can be used as a first
+complete R code examples were still needed. Given the R community's interest in 
+data analysis, data science, and machine learning, and the broad support of R packages for data mining, there was a noticeable gap that was filled by this learning resource.
+This resource targets advanced undergraduate and graduate students and can be used as a 
+component for a first
 introduction to data mining. 
 
 # Learning Objectives and Content
@@ -50,30 +51,30 @@ learn how to
 * prepare and understand data,
 * perform classification,
 * perform association analysis, and
-* perform cluster analysis
+* perform cluster analysis.
 
-using self-contained and annotated R code examples that work with small datasets
+The method is to present self-contained and annotated R code examples that work with small datasets
 carefully chosen to show the learner important aspects of data mining. 
-The learner can copy-and-paste the examples into an R markdown notebook to experiment with 
+The learner can copy and paste the examples into an R markdown notebook to experiment with 
 the code and the provided example data and then modify the code to work on their own data sets.
 This learning-by-doing approach has worked very well in preparing students to work
 with more complex real-world datasets by relieving them from dealing with 
-too many implementation details while exploring the concepts. 
+too many low-level implementation details while exploring the concepts. 
 
-To make using the resource easier, the resources mirrors 
-the structure of the textbook. After a short introduction, 
-Chapter 2 discusses data types in R, data quality concerns and data preprocessing.
-In addition data exploration and visualization examples are given. Chapters 3 and 4
+The resource mirrors textbook's structure so it can easily used along 
+with the textbook. After a short introduction, 
+Chapter 2 discusses data types in R, data quality concerns ,and data preprocessing.
+In addition, data exploration and visualization examples are given. Chapters 3 and 4
 cover classification methods, model selection, model evaluation, different 
-types of classifiers, and issues like class imbalance. Chapter 5 introduces
+types of classifiers, and essential practical issues like class imbalance. Chapter 5 introduces
 association analysis with a strong emphasis on visualization. Chapter 7 presents 
-examples for cluster analysis including popular algorithms, cluster evaluation, and
+examples of cluster analysis, including popular algorithms, cluster evaluation, and
 the effect of outliers.
 
 # Instructional Design
 
 This resource does not replace the Introduction to Data Mining textbook or instruction 
[…]
