+++ paper/paper.md
@@ -32,9 +32,10 @@ The materials have been made publicly available at: <https://mhahsler.github.io/
 
 # Statement of Need
 
-The textbook Introduction to Data Mining has been one of the most popular choices
-to learn data mining concepts.
-Several of the key chapters have been made available for free by the authors [@Tan2018].
+The textbook Introduction to Data Mining [@Tan2018] has been one of the most popular choices
+to learn and teach data mining concepts.
+Some of the most important chapters have been made available for free by the authors on
+the [books's website](https://www-users.cse.umn.edu/~kumar001/dmbook/index.php).
 One of the authors also provides Python Jupyter notebooks with examples, but
 complete R code examples were still needed. Given the R community's interest in 
 data analysis, data science, and machine learning, and the broad support of R packages for data mining, there was a noticeable gap that was filled by this learning resource.
@@ -53,17 +54,18 @@ learn how to
 * perform association analysis, and
 * perform cluster analysis.
 
-The method is to present self-contained and annotated R code examples that work with small datasets
-carefully chosen to show the learner important aspects of data mining. 
-The learner can copy and paste the examples into an R markdown notebook to experiment with 
-the code and the provided example data and then modify the code to work on their own data sets.
-This learning-by-doing approach has worked very well in preparing students to work
-with more complex real-world datasets by relieving them from dealing with 
+The resource presents self-contained and annotated R code examples that work with small datasets
+carefully chosen to show the learner many important aspects of data mining. 
+The learner can copy and paste the examples into a new R markdown notebook to experiment with 
+the code and the provided example data. Small exercises encourage the learner to
+modify the code by applying it to a different dataset.
+This learning-by-doing approach has worked well in preparing students to work
+with more complex real-world datasets by initially relieving them from dealing with 
 too many low-level implementation details while exploring the concepts. 
 
-The resource mirrors textbook's structure so it can easily used along 
+The resource mirrors the textbook's structure so it can be easily used along 
 with the textbook. After a short introduction, 
-Chapter 2 discusses data types in R, data quality concerns ,and data preprocessing.
+Chapter 2 discusses data types in R, data quality concerns, and data preprocessing.
 In addition, data exploration and visualization examples are given. Chapters 3 and 4
 cover classification methods, model selection, model evaluation, different 
 types of classifiers, and essential practical issues like class imbalance. Chapter 5 introduces
@@ -75,7 +77,7 @@ the effect of outliers.
 
 This resource does not replace the Introduction to Data Mining textbook or instruction 
 by a teacher, it instead provides supporting material for 
-learning to implement data mining concepts.
+learning to implement data mining concepts in R.
 The learner is expected to have some programming experience and basic 
 statistics knowledge. 
 
@@ -93,10 +95,10 @@ between the slides and the code examples is provided by the R symbol on the slid
 where example code is available. 
 The code examples can be assigned to be studied by the students outside of class
 or used by the instructor in class.
+
 Designing assignments and assessments is left to the instructor since they 
 depend on the level and field of study of the students (e.g., computer science, 
 statistic, economics, or business).
-
 For example, for undergraduates, it is suggested to ask the students
 to apply the data mining techniques to a small, clean instructional data set 
 (sample exercises are available in the resource at the end of each chapter), 
@@ -105,7 +107,7 @@ significant amount of cleaning and preprocessing.
 
 # Story of the Project
 
-Since starting to teach data mining with R in the Spring 2013 years, I have been developing 
+Since starting to teach data mining with R in the Spring 2013, I have been developing 
 the Companion for Introduction to Data Mining
 resource mainly based on caret [@caret2008], and a set of packages
 developed with students to better support different data mining tasks (e.g., 
[…]
