### 2024-06-24 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @mhahsler
@stats-tgeorge

Dear reviewers and editor,

Thank you for the very useful comments! My responses are typeset as blockquotes.

I was initially under the impression that there was a strict 2-page limit, including the references. However, it looks like the limit is 1000 words, so I was able to provide more information in the paper to address all the comments.

Some general changes are:
* I have rearranged some contents to follow the 2nd edition of the textbook more closely. The section numbers match now up to level 2. This will make it easier to match the theory in the textbook with the code in the resource.
* I added exercises to all chapters.
* I cleaned the code and updated the deprecated use of dplyr functions.

Responses to the individual comments:

- [x] Statement of need: Please specify the target audience. Consider different levels of courses the materials could be considered for. A first course in statistics or data science? A senior undergraduate course? A graduate course? Suggested pre-requisites?

> In the "Statement of Need," I now specify, "This resource targets advanced undergraduate and graduate students and can be used as a first introduction to data mining." I believe there are better sources out there for statistics and data science courses.

- [x] Usage: In your JOSE paper, discuss how instructions can use all or parts of this for teaching. You discuss how self-learners might use it but not how someone could adopt all or parts of the materials into their class.

> I have rewritten the "Instructional Design" section to focus more on how an instructor would integrate it into a course.

**Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)**

- [x] Learning objectives: You give a brief outline in Chapter 1 of the topics to be covered in your companion book. I would like to see more depth. Is this going to be application-focused? Or, will both the theory and application be presented? Consider what would be useful to see to determine if someone should keep reading. Many books have these topics, what will they find about them, and how to implement them with your companion?

> I added a new section called "Learning Objectives and Content," which now provides more details on the content.

- [x] Content quality:
  - Chapter 4 is missing any description of the outputs of the classification models. Specifically 4.3.4, 4.3.5, 4.3.6, 4.3.7, 4.3.8,
  - The same is true in 4.7, with no comments. There are also many conditions that need to be met in order to use many of these approaches.

> When I reorganized Chapter 4, I added an explanation of how to read the models' output and a baseline model for the section Model Comparison.

  - In 5.5.3 there is a coding warning but it is not discussed and appears to be important.

> The explanation for the warning was not in the right place. I have moved it to right after the warning ocurrs.

  - In 5.7 the interactive table is overlapping with your table of contents on the right.

> This is now resolved.

**JOSE Paper**

- [x] Statment of need: Same as above.

> See response above.

- [x] Description: Please add more comments on the organization. You have the online book and webpage with a table of materials people can click on as wanted. Give an outline of the materials available, and what each resource is for.

> I have added a short list in section "Learning Objectives and Content".

- [x] References: DOI - see the mentioned failed DOIs in the editorial bot response on the review.

> I have added one missing DOI.
> The textbook Introduction to Data Mining has no DOI.

- [x] Story - add a bit of the background story from your book into the JOSE paper

> I have added a "Story of the Project" section to the paper.

General Comments (do not have to be addressed)
- Consider an easier introduction to the dplyr/tidyverse verbs. Jumping directly to using something like across() within a summarize before showing examples of one or the other will be challenging for readers.

> I have added `across()` with an example and an explanation to Chapter 1.

- Consider including _Palmer Penguins_ to diversity your examples on unsupervised techniques (or supervised with grouping).

> Thank you for bringing this very well-made dataset to my attention. I use it now through the resources for exercises (see the Exercises sections at the end of each Chapter).

- I would rename your 2.11.2 section, grouped operations or calculations.

> Good idea. I have renamed the section.

- As nice as it is to put your packages all in 1 place, you might consider putting packages with very specific uses in those specific sections. This will allow easier copy/paste/use.

> I went through the chapters and added `library()` calls to the subsections where a package is used. The code in each chapter is now self-contained, and most sections can be run as long as tidyverse and the data set is loaded.

- I'm assuming your ordering is due to the book ordering. If not, the order should be 2.11 -> 2.12 both before 2.7 or 2.8.

> Yes, this ordering is due to the textbook. Changing the order would make it more confusing for students following the textbook.

- In future versions, _tidymodels_ has come a long way and can help to unify much of the code under more similar syntax. The author of Caret, Max Kuhn, is now also working on _tidymodels_.  I see you gave that as a future reference.

> This is true. tidymodels is on my list to add in the future, and I will probably produce a second version with it at some point.
