+++ book_src/_common.R
@@ -17,6 +17,6 @@ local({
 
 # example chunk options set globally
 knitr::opts_chunk$set(
-  comment = "#>",
+#  comment = "#>",
   collapse = TRUE
   )
