+++ book/reference-keys.txt
@@ -1,6 +1,16 @@
 introduction
 used-software
+base-r
+vectors
+operations
+functions
+subsetting-vectors
+strings
+plotting
+objects
 tidyverse
+tibbles
+transformations
 ggplot2
 data
 types-of-data
