+++ book_src/01_intro.Rmd
@@ -15,33 +15,25 @@ annotated R code examples to complement the textbook. To make following
 along easier, we follow the chapters in data mining textbook which is
 organized by the main data mining tasks:
 
--   
+2.  [Data] covers types of data and also includes data preparation and
+    exploratory data analysis in the chapter.
 
-    2.  [Data] covers types of data and also includes data preparation
-        and exploratory data analysis in the chapter.
+3.  [Classification: Basic Concepts and Techniques] introduces decision
+    trees, model training and evaluation.
 
--   
-
-    3.  [Classification: Basic Concepts and Techniques] introduces
-        decision trees, model training and evaluation.
-
--   
-
-    4.  [Classification: Alternative techniques] introduces and compares
+4.  [Classification: Alternative techniques] introduces and compares
         methods including rule-based classifiers, nearest neighbor
         classifiers, naive Bayes classifier, logistic regression and
         artificial neural networks.
 
--   
-
-    5.  [Association Analysis: Basic Concepts and Algorithms] covers
+5.  [Association Analysis: Basic Concepts and Algorithms] covers
         frequent itemset and association rule generation and analysis
         including visualization. Some code covers also topics discussed
         in the textbook's Chapter 6.
 
--   
+6.  Unused 
 
-    7.  [Cluster Analysis] discusses clustering approaches including
+7.  [Cluster Analysis] discusses clustering approaches including
         k-means, hierarchical clustering, DBSCAN and how to evaluate
         clustering results.
 
@@ -50,25 +42,23 @@ For completeness, we have added chapters on [Regression\*] and on
 contain code examples for methods that are not described in the data
 mining textbook.
 
+This book assumes that you are
+familiar with the basics of R, how to run R code, and install packages.
+The rest of this chapter will provide an overview and point you to where 
+you can learn more about R and the used packages.
+
 ## Used Software
 
-This companion book assumes that you have
+To use this book you need to have
 [R](https://cran.r-project.org/) and [RStudio
-Desktop](https://www.rstudio.com/products/rstudio/) installed and are
-familiar with the basics of R, how to run R code, and install packages.
-If you are new to R, working through the official R manual [An
-Introduction to
-R](https://cran.r-project.org/doc/manuals/r-release/R-intro.html)
-[@RCore2021] will get you started. There are many introduction videos
-for RStudio available, and a basic video that shows how to run code and
-how to install packages will suffice.
+Desktop](https://www.rstudio.com/products/rstudio/) installed.
 
 Each book chapter will use a set of packages that must be installed. The
 installation code can be found at the beginning of each chapter. Here is
 the code to install the packages used in this chapter:
 
 ```{r setup_01, message=FALSE}
-pkgs <- c('tidyverse', 'ggplot2')
+pkgs <- c('tidyverse')
 
 pkgs_install <- pkgs[!(pkgs %in% installed.packages()[,"Package"])]
 if(length(pkgs_install)) install.packages(pkgs_install)
@@ -80,30 +70,27 @@ source("format_pkgs.R")
 all_pkgs <- union(all_pkgs, pkgs)
 ```
 
-The packages used for this chapter are: `r format_pkgs(pkgs)`
-
-The code examples in this book uses the R packages `tidyverse` to manipulate 
-data and `ggplot2` for visualization. These packages make working with
-data in R very convenient. 
-A great introduction to these useful tools can be
-found in the freely available web book [R for Data
-Science](https://r4ds.hadley.nz/) by @Wickham2023.
-We will give a few examples later in this chapter.
-
-Everything we do in R is built on top of the code R programming language
-and the packages that are automatically installed with R. This is typically 
[…]
