+++ book_src/packages.bib
@@ -9,9 +9,9 @@
 @Manual{R-arulesViz,
   title = {arulesViz: Visualizing Association Rules and Frequent Itemsets},
   author = {Michael Hahsler},
-  year = {2024},
-  note = {R package version 1.5.3},
+  note = {R package version 1.5.3-1},
   url = {https://github.com/mhahsler/arulesViz},
+  year = {2024},
 }
 
 @Manual{R-base,
