### 2024-06-29 505cf34
Added chapter 6 (advanced association rule mining).
