+++ book_src/06_arules.Rmd
@@ -1,4 +1,249 @@
 # Association Analysis: Advanced Concepts
 
-No code is available for this Chapter. Some topics were already covered in the 
-code for the previous chapter.
+This chapter discusses a few advanced concepts of association analysis.
+First, we look at how categorical and continuous attributes are converted
+into items. The we look at integrating item hierarchies into the analysis.
+Finally, sequence pattern mining is introduced.
+
+### Packages Used in this Chapter {.unnumbered}
+
+```{r setup_06}
+pkgs <- c("arules", "arulesSequences", "tidyverse")
+
+pkgs_install <- pkgs[!(pkgs %in% installed.packages()[,"Package"])]
+if(length(pkgs_install)) install.packages(pkgs_install)
+```
+
+```{r setup_06-2, include=FALSE}
+source("_common.R")
+source("format_pkgs.R")
+all_pkgs <- union(all_pkgs, pkgs)
+```
+
+The packages used for this chapter are: `r format_pkgs(pkgs)`
+
+## Handling Categorical Attributes
+
+Categorical attributes are nominal or ordinal variables. 
+In R they are `factors` or `ordinal`. They are
+translated into a series of binary items (one for each level constructed as ⁠variable
+`name = level`⁠). Items cannot represent order and this ordered factors lose
+the order information. Note that nominal variables need to be encoded as 
+factors (and not characters or numbers) before converting them into transactions.
+
+For the special case of Boolean variables (`logical`), the `TRUE` value is 
+converted into an item with the name of the variable and for the `FALSE` 
+values no item is created.
+
+We will give an example in the next section.
+
+## Handling Continuous Attributes
+
+Continuous variables cannot directly be represented as items and need to be 
+discretized first (see [Discretization] in Chapter 2). 
+An item resulting from discretization might be `age>18` and 
+the column contains only `TRUE` or `FALSE`. Alternatively, it can be a factor 
+with levels `age<=18`, `⁠50=>age>18`⁠ and `age>50`. These will be automatically
+converted into 3 items, one for each level. Discretization is described in functions 
+`discretize()` and `discretizeDF()` to discretize all columns in a data.frame.
+
+We give a short example using the iris dataset. We add an extra `logical` column
+to show how Boolean attributes are converted in to items.
+
+```{r }
+data(iris)
+
+## add a Boolean attribute
+iris$Versicolor <- iris$Species == "versicolor"
+head(iris)
+```
+
+The first step is to
+discretize continuous attributes (marked as `<dbl>` in the table above).
+We discretize the two Petal features.
+
+```{r }
+library(tidyverse)
+library(arules)
+
+iris_disc <- iris %>% 
+  mutate(Petal.Length = discretize(Petal.Length, 
+                                   method = "frequency", breaks = 3, 
+                                   labels = c("short", "medium", "long")),
+         Petal.Width = discretize(Petal.Width,
+                                  method = "frequency", breaks = 2, 
+                                  labels = c("narrow", "wide"))
+         )
+  
+
+head(iris_disc)
+```
+
+Next, we convert the dataset into transactions.
+
+```{r }
+trans <- transactions(iris_disc)
+trans
+```
+
+The conversion creates a warning because there are still two undiscretized columns
+in the data. The warning indicates that the default discretization is used 
+automatically.
+
+```{r }
+itemLabels(trans)
+```
+
+We see that all continuous variables are discretized and the different ranges 
+create an item. For example `Petal.Width` has the two items `Petal.Width=narrow`
+and `Petal.Width=wide`. The automatically discretized variables show intervals.
+`Sepal.Length=[4.3,5.4)` means that this item used for flowers with 
+a sepal length between 4.3 and 5.4 cm.
+
+The species is converted into three items, one for each class. The logical 
[…]
