### 2024-07-02 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @stats-tgeorge
@mhahsler

Looking good! There is an inconsistency in your Statements of Need.  You have them on your website, your repo readme, and your JOSE paper and they do not all match. I believe your JOSE paper version is your most updated. Can you update the other locations?
