### 2024-07-02 #223 @stats-tgeorge
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.18637/jss.v014.i15 is OK
> - 10.21105/joss.01686 is OK
> - 10.18637/jss.v091.i01 is OK
> - 10.18637/jss.v025.i03 is OK
> - 10.32614/RJ-2017-047 is OK
> - 10.18637/jss.v028.i05 is OK
> - 10.32614/CRAN.package.tidymodels is OK
>
> MISSING DOIs
>
> - 10.1002/0471687545.ch1 may be a valid DOI for title: Introduction to Data Mining
>
> INVALID DOIs
>
> - None
> ```

@mhahsler Looks like there is still a DOI issue.
