### 2024-07-03 #223 @mhahsler
@stats-tgeorge The book "Introduction to Data Mining" seems to not have a registered DOI. The DOI in the check is for an unrelated book chapter with the same name.

It has a DOI used internally only by ACM, but it is not registered with DOI.org
https://dl.acm.org/doi/10.5555/3208440
