+++ paper/paper.md
@@ -17,115 +17,108 @@ affiliations:
 
 # Summary
 
-An R Companion for Introduction to Data Mining is an open-source learning and 
+An R Companion for Introduction to Data Mining is an open-source learning and
 teaching resource that covers how to implement data mining concepts using R.
-It is designed to accompany the popular
-data mining textbook _Introduction to Data Mining_ [@Tan2018]
-to study the implementation of the basic data mining concepts 
-including data preparation, classification,
-clustering, and association analysis. 
-The resource uses complete, annotated examples 
-to demonstrate how data mining concepts can be translated into R code.
+It is designed to accompany the popular data mining textbook _Introduction to
+Data Mining_ [@Tan2018] to study the implementation of the basic data mining
+concepts including data preparation, classification, clustering, and
+association analysis.  The resource uses complete, annotated examples to
+demonstrate how data mining concepts can be translated into R code.
 
 
 The materials have been made publicly available at: <https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/> and licensed under the [Creative Commons Attribution-NonCommercial-ShareAlike 4.0 (CC BY-NC-SA 4.0) License](https://creativecommons.org/licenses/by-nc-sa/4.0/).
 
 # Statement of Need
 
-The textbook Introduction to Data Mining [@Tan2018] has been one of the most popular choices
-to learn and teach data mining concepts.
-Some of the most important chapters have been made available for free by the authors on
-the [books's website](https://www-users.cse.umn.edu/~kumar001/dmbook/index.php).
+The textbook Introduction to Data Mining [@Tan2018] has been one of the most
+popular choices to learn and teach data mining concepts.  Some of the most
+important chapters have been made available for free by the authors on the
+[books's website](https://www-users.cse.umn.edu/~kumar001/dmbook/index.php).
 One of the authors also provides Python Jupyter notebooks with examples, but
-complete R code examples were still needed. Given the R community's interest in 
-data analysis, data science, and machine learning, and the broad support of R packages for data mining, there was a noticeable gap that was filled by this learning resource.
-This resource targets advanced undergraduate and graduate students and can be used as a 
-component for a first
-introduction to data mining. 
+complete R code examples were still needed. Given the R community's interest in
+data analysis, data science, and machine learning, and the broad support of R
+packages for data mining, there was a noticeable gap that was filled by this
+learning resource.  This resource targets advanced undergraduate and graduate
+students and can be used as a component for a first introduction to data
+mining. 
 
 # Learning Objectives and Content
 
-The resource assumes basic knowledge of programming and statistics.
-The learning objectives are to
-learn how to
+The resource assumes basic knowledge of programming and statistics.  The
+learning objectives are to learn how to
 
 * prepare and understand data,
 * perform classification,
 * perform association analysis, and
 * perform cluster analysis.
 
-The resource presents self-contained and annotated R code examples that work with small datasets
-carefully chosen to show the learner many important aspects of data mining. 
-The learner can copy and paste the examples into a new R markdown notebook to experiment with 
-the code and the provided example data. Small exercises encourage the learner to
-modify the code by applying it to a different dataset.
-This learning-by-doing approach has worked well in preparing students to work
-with more complex real-world datasets by initially relieving them from dealing with 
-too many low-level implementation details while exploring the concepts. 
+The resource presents self-contained and annotated R code examples that work
+with small datasets carefully chosen to show the learner many important aspects
+of data mining. The learner can copy and paste the examples into a new R
+markdown notebook to experiment with the code and the provided example data.
+Small exercises encourage the learner to modify the code by applying it to a
+different dataset. This learning-by-doing approach has worked well in
+preparing students to work with more complex real-world datasets by initially
+relieving them from dealing with too many low-level implementation details
[…]
