### 2024-07-03 207470b
removed bibtex ref.
