### 2024-07-03 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @mhahsler
@stats-tgeorge thank you for catching this. I have updated the statement in the README on GitHub.

https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples

Is the version on my website, the github.io verions that is linked from my homepage? It automatically updates GitHub with changes.

https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/
