+++ book_src/03_classification_basics.Rmd
@@ -7,7 +7,11 @@ editor_options:
 # Classification: Basic Concepts and Techniques
 
 This chapter introduces decision trees for classification and discusses
-how models are built and evaluated.
+how models are built and evaluated. 
+
+The corresponding chapter of 
+the data mining textbook is available online: 
+[Chapter 3: Classification: Basic Concepts and Techniques.](https://www-users.cs.umn.edu/~kumar001/dmbook/ch3_classification.pdf)
 
 ### Packages Used in this Chapter {.unnumbered}
 
