### 2024-08-09 cd68618
Added R wizard GPT.
