### 2024-08-11 16716b6
added group_by and updated stratified sampling using dplyr.
