+++ book_src/packages.bib
@@ -65,12 +65,12 @@ al.},
 }
 
 @Manual{R-dbscan,
-  title = {dbscan: Density-Based Spatial Clustering of Applications with Noise (DBSCAN)
-and Related Algorithms},
+  title = {dbscan: Density-Based Spatial Clustering of Applications with Noise
+(DBSCAN) and Related Algorithms},
   author = {Michael Hahsler and Matthew Piekenbrock},
-  year = {2024},
-  note = {R package version 1.2-0},
+  note = {R package version 1.2-0-1},
   url = {https://github.com/mhahsler/dbscan},
+  year = {2024},
 }
 
 @Manual{R-dplyr,
@@ -212,9 +212,9 @@ Theory Group (Formerly: E1071), TU Wien},
 
 @Manual{R-MASS,
   title = {MASS: Support Functions and Datasets for Venables and Ripley's MASS},
-  author = {Brian Ripley},
-  year = {2023},
-  note = {R package version 7.3-60},
+  author = {Brian Ripley and Bill Venables},
+  year = {2024},
+  note = {R package version 7.3-61},
   url = {http://www.stats.ox.ac.uk/pub/MASS4/},
 }
 
