### 2024-08-11 2cc150e
added link to book.
