+++ book_src/01_intro.Rmd
@@ -38,8 +38,8 @@ organized by the main data mining tasks:
         k-means, hierarchical clustering, DBSCAN and how to evaluate
         clustering results.
 
-For completeness, we have added chapters on [Regression\*] and on
-[Logistic Regression\*]. Sections in this book followed by an asterisk
+For completeness, we have added chapters on [Regression*] and on
+[Logistic Regression*]. Sections in this book followed by an asterisk
 contain code examples for methods that are not described in the data
 mining textbook.
 
