+++ book_src/02_data.Rmd
@@ -283,16 +283,18 @@ More sophisticated sampling procedures are implemented in the package
 ### Dimensionality Reduction
 
 The number of features is often called the dimensional of the data following the 
-idea that each feature (at least the numeric features) can be seen as an axis of the data.
+idea that each feature (at least the numeric features) can be seen as an axis of the data. High-dimensional data is harder to analyze by the user (e.g., visualize). 
+It also is problematic for many data mining algorithms since it requires more memory 
+and computational resources. 
 
-Common feature preprocessing includes 
-[dimensionality reduction](https://en.wikipedia.org/wiki/Dimensionality_reduction)
-which tries to represent high-dimensional data in a low-dimensional space so that the 
-low-dimensional representation retains some meaningful properties (e.g., information about
-similarity or distances) of the original data. A special case is feature selection.
+[Dimensionality reduction](https://en.wikipedia.org/wiki/Dimensionality_reduction)
+ tries to represent high-dimensional data in a low-dimensional space so that the 
+low-dimensional representation retains some meaningful properties (e.g., information about similarity or distances) of the original data. Dimensionality reduction is 
+used for visualization and as a prepossessing technique before using data mining 
+models.
 
-Other important feature preprocessing includes discretization and feature scaling.
-
+A special case of dimensionality reduction is feature selection (see 
+[Feature Selection*] in Chapter 3).
 
 #### Principal Components Analysis (PCA)
 
@@ -453,7 +455,20 @@ to the result of the projection using PCA.
 
 Non-parametric multidimensional scaling performs MDS while relaxing the
 need of linear relationships. Methods are available in package `MASS` as
-functions `isoMDS()` and `sammon()`.
+functions `isoMDS()` (implements [isoMAP](https://en.wikipedia.org/wiki/Isomap)) and `sammon()`.
+
+#### Other Nonlinear Dimensionality Reduction Methods
+
+[Nonlinear dimensionality reduction](https://en.wikipedia.org/wiki/Nonlinear_dimensionality_reduction)
+is also called manifold learning or creating a low-Dimensional embedding.
+Many methods are available with varying properties exist. 
+Some popular methods are:
+
+* [t-distributed stochastic neighbor embedding](https://en.wikipedia.org/wiki/T-distributed_stochastic_neighbor_embedding) (`Rtsne()` in package `Rtsne`) and uniform manifold approximation and projection
+(`umap()` in package `umap`) are used 
+for projecting data to 2 dimensions for visualization.
+* [Word2vec](https://en.wikipedia.org/wiki/Word2vec) (`word2vec` in package (`word2vec`)) is used in [natural language processing](https://cran.r-project.org/web/views/NaturalLanguageProcessing.html) to convert words to numeric 
+vectors for similarity calculation.
 
 ### Feature Subset Selection
 
