### 2024-08-15 #223 @stats-tgeorge
> @mhahsler checklist is complete

Thank you so much for getting back to this. This one was a big review and I appreciate your time!
