### 2024-08-15 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @stats-tgeorge
@mhahsler in your paper references ([bib file](https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples/blob/master/paper/paper.bib)), the Kuhn & Max reference is formatted differently. Please fix this.
