+++ paper/paper.bib
@@ -83,9 +83,9 @@
   doi = {10.18637/jss.v028.i05},
   number = {5},
   journal = {Journal of Statistical Software},
-  author = {{Kuhn} and {Max}},
+  author={Kuhn, Max},
   year = {2008},
-  pages = {1–26},
+  pages = {1-26}
 }
 
 @Manual{tidymodels2020,
