### 2024-08-15 mhahsler/Introduction_to_Data_Mining_R_Examples#6 @mhahsler
Interesting. The author field seems to be broken in the caret package.

```
> citation("caret")
To cite caret in publications use:

  Kuhn, M. (2008). Building Predictive Models in R Using the caret Package. Journal of Statistical Software,
  28(5), 1–26. https://doi.org/10.18637/jss.v028.i05

A BibTeX entry for LaTeX users is

  @Article{,
    title = {Building Predictive Models in R Using the caret Package},
    volume = {28},
    url = {https://www.jstatsoft.org/index.php/jss/article/view/v028i05},
    doi = {10.18637/jss.v028.i05},
    number = {5},
    journal = {Journal of Statistical Software},
    author = {{Kuhn} and {Max}},
    year = {2008},
    pages = {1–26},
  }
```

I have fixed the author field in the bibtex file and will let Max know.
