### 2024-08-15 #223 @mhahsler
@stats-tgeorge Additional Author Tasks After Review is Complete

Version Number: 1.0.1 (https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples/releases/tag/1.0.1)
Archive: figshare DOI: [10.6084/m9.figshare.26750404.v1](https://doi.org/10.6084/m9.figshare.26750404.v1)
Note on license: figshare only gives a choice for CC BY 4.0, while the book uses CC BY-NC 4.0. I can change everything to  CC BY 4.0, if that is necessary.
