### 2024-08-15 #223 @mhahsler
> **License info**:
>
> 🔴 License found: `Creative Commons Attribution Share Alike 4.0 International` (Not [OSI approved](https://opensource.org/licenses))

Does this mean that I cannot use CC?
