### 2024-08-15 #223 @stats-tgeorge
@mhahsler the DOI needs updating on the JOSE paper. Once that is fixed I will recommend accepting.
