### 2024-08-16 #223 @stats-tgeorge
> @stats-tgeorge That is geat news! Thanks for helping me through this process.
>
>
>
> About the DOI: I assume you are asking for this missing DOI
>
>
>
>
>
> ```
>
> MISSING DOIs
>
>
>
> - 10.1002/0471687545.ch1 may be a valid DOI for title: Introduction to Data Mining
>
> ```
>
> The referenced book is Tan, P.-N., Steinbach, M. S., Karpatne, A., & Kumar, V. (2017). _Introduction to data
>
> mining (2nd Edition)._ Pearson. ISBN: 978-0133128901
>
>
>
> The suggested DOI is for a different book: _Discovering Knowledge in Data: An Introduction to Data Mining_
>
> by Daniel T. Larose Ph.D.
>
>
>
>
>
> It seems like not all books have assigned DOIs. There is no DOI on the publisher's page
>
> https://www.pearson.com/en-us/subject-catalog/p/introduction-to-data-mining/P200000003204/9780137506286
>
>
>
> I also cannot find one registered with crossref
>
> https://search.crossref.org/search/works?q=%22Introduction+to+Data+Mining%22+Tan+Steinbach+Kumar&from_ui=yes
>
>
>
> As far as I can tell, the book does not have a DOI at this point.

Good morning! It's what I'm here for - I am sorry this took so long!

I saw you mentioned that DOI before. I am referring to the DOI on the left side of the rendered JOSE paper. Looks like you had a placeholder there.
