### 2024-08-19 134c941
Better control for upper/lowercase.
