+++ paper/paper.bib
@@ -14,7 +14,7 @@
 
 @Article{arules2005,
   AUTHOR       = {Michael Hahsler and Bettina Gr{\"u}n and Kurt Hornik},
-  TITLE        = {arules -- {A} Computational Environment for Mining
+  TITLE        = {{arules} -- {A} Computational Environment for Mining
                   Association Rules and Frequent Item Sets},
   JOURNAL      = {Journal of Statistical Software},
   YEAR         = 2005,
@@ -63,7 +63,7 @@
 }
 
 @Article{arulesViz2017,
-  title = {arules{V}iz: {I}nteractive Visualization of Association Rules with {R}},
+  title = {{arulesViz}: {I}nteractive Visualization of Association Rules with {R}},
   author = {Michael Hahsler},
   year = {2017},
   journal = {R Journal},
