+++ book_src/04_classification_alternative.Rmd
@@ -31,7 +31,9 @@ all_pkgs <- union(all_pkgs, pkgs)
 set.seed(1234)
 ```
 
-The packages used for this chapter are: `r format_pkgs(pkgs)`
+The packages used for this chapter are: 
+
+`r format_pkgs(pkgs)`
 
 
 ## Types of Classifiers
@@ -95,7 +97,9 @@ We will use for all models the same validation sets by
 creating a fixed sampling scheme (10-folds). This will help when we can compare the fitted models later.
 
 ```{r }
-train_index <- createFolds(Zoo_train$type, k = 10, returnTrain = TRUE)
+train_index <- createFolds(Zoo_train$type, 
+                           k = 10, 
+                           returnTrain = TRUE)
 ```
 
 The fixed folds are used in `train()` with the argument
@@ -121,7 +125,8 @@ rulesFit <- Zoo_train |> train(type ~ .,
   method = "PART",
   data = _,
   tuneLength = 5,
-  trControl = trainControl(method = "cv", index = train_index))
+  trControl = trainControl(method = "cv", 
+                           index = train_index))
 rulesFit
 ```
 The model selection results are shown in the table. This is the selected model:
@@ -152,7 +157,8 @@ knnFit <- Zoo_train |> train(type ~ .,
   data = _,
   preProcess = "scale",
   tuneGrid = data.frame(k = c(1, 3, 5, 7, 9)),
-	trControl = trainControl(method = "cv", index = train_index))
+	trControl = trainControl(method = "cv", 
+	                         index = train_index))
 knnFit
 ```
 
@@ -167,7 +173,7 @@ keep the complete dataset. This is why the final model just gives us a summary s
 ## Naive Bayes Classifier
 
 Caret's train formula interface translates logicals and factors into dummy 
-variables which the classifier interprets as numbers so it would used a Gaussian naive Bayes estimation. To avoid this, I directly specify `x` and `y`.
+variables which the classifier interprets as numbers so it would used a [Gaussian naive Bayes](https://en.wikipedia.org/wiki/Naive_Bayes_classifier#Gaussian_naive_Bayes) estimation. To avoid this, I directly specify `x` and `y`.
 
 ```{r }
 NBFit <- train(x = as.data.frame(Zoo_train[, -ncol(Zoo_train)]), 
@@ -175,7 +181,8 @@ NBFit <- train(x = as.data.frame(Zoo_train[, -ncol(Zoo_train)]),
                method = "nb",
                tuneGrid = data.frame(fL = c(.2, .5, 1, 5), 
                                      usekernel = TRUE, adjust = 1),
-               trControl = trainControl(method = "cv", index = train_index))
+               trControl = trainControl(method = "cv", 
+                                        index = train_index))
 NBFit
 ```
 
@@ -221,7 +228,8 @@ logRegFit <- Zoo_train |> train(type ~ .,
   method = "multinom",
   data = _,
   trace = FALSE, # suppress some output
-	trControl = trainControl(method = "cv", index = train_index))
+	trControl = trainControl(method = "cv", 
+	                         index = train_index))
 logRegFit
 ```
 
@@ -244,7 +252,8 @@ nnetFit <- Zoo_train |> train(type ~ .,
   method = "nnet",
   data = _,
 	tuneLength = 5,
-	trControl = trainControl(method = "cv", index = train_index),
+	trControl = trainControl(method = "cv", 
+	                         index = train_index),
   trace = FALSE # no progress output
   )
 nnetFit
@@ -326,9 +335,13 @@ reduce overfitting.
 ```{r message=FALSE, warning=FALSE}
 model <- keras_model_sequential(input_shape = c(ncol(X)), 
[…]
