+++ book_src/03_classification_basics.Rmd
@@ -4,7 +4,7 @@ editor_options:
     wrap: 72
 ---
 
-# Classification: Basic Concepts and Techniques
+# Classification: Basic Concepts
 
 This chapter introduces decision trees for classification and discusses
 how models are built and evaluated. 
@@ -17,14 +17,16 @@ the data mining textbook is available online:
 
 ```{r setup_03}
 pkgs <- c("caret", "FSelector", "lattice", "mlbench", 
-          "palmerpenguins", "party", "pROC", "rpart", "rpart.plot", 
-          "sampling", "tidyverse")
+          "palmerpenguins", "party", "pROC", "rpart", 
+          "rpart.plot", "tidyverse")
 
 pkgs_install <- pkgs[!(pkgs %in% installed.packages()[,"Package"])]
 if(length(pkgs_install)) install.packages(pkgs_install)
 ```
 
-The packages used for this chapter are: `r format_pkgs(pkgs)`
+The packages used for this chapter are: 
+
+`r format_pkgs(pkgs)`
 
 ```{r setup_03-2, include=FALSE}
 source("_common.R")
@@ -137,7 +139,7 @@ the data is ready for model learning.
 ```{r }
 Zoo <- Zoo |>
   mutate(across(where(is.logical), 
-                function (x) factor(x, levels = c(TRUE, FALSE)))) |>
+         function (x) factor(x, levels = c(TRUE, FALSE)))) |>
   mutate(across(where(is.character), factor))
 
 summary(Zoo)
@@ -169,9 +171,8 @@ tree_default
   the first argument of `rpart`, the syntax `data = _` is used to specify
   where the data in `Zoo` goes. The call is equivalent to
   `tree_default <- rpart(type ~ ., data = Zoo)`. 
-
-- The formula models the`type` variable by all other features represented by `.`.
-
+- The formula models the `type` variable by all other features represented by 
+  a single period (`.`).
 - The class variable needs to be a factor to be recognized as nominal 
   or rpart will create a regression tree instead of a decision tree. 
   Use `as.factor()` on the column with the class label first, if necessary.
@@ -194,9 +195,9 @@ I will make up my own animal: A lion with feathered wings.
 ```{r }
 my_animal <- tibble(hair = TRUE, feathers = TRUE, eggs = FALSE,
   milk = TRUE, airborne = TRUE, aquatic = FALSE, predator = TRUE,
-  toothed = TRUE, backbone = TRUE, breathes = TRUE, venomous = FALSE,
-  fins = FALSE, legs = 4, tail = TRUE, domestic = FALSE,
-  catsize = FALSE, type = NA)
+  toothed = TRUE, backbone = TRUE, breathes = TRUE, 
+  venomous = FALSE, fins = FALSE, legs = 4, tail = TRUE, 
+  domestic = FALSE, catsize = FALSE, type = NA)
 ```
 
 The data types need to match the original data so we
@@ -293,8 +294,8 @@ tree_full <- Zoo |>
         control = rpart.control(minsplit = 2, cp = 0))
 rpart.plot(tree_full, extra = 2, 
            roundint=FALSE,
-            box.palette = list("Gy", "Gn", "Bu", "Bn", 
-                               "Or", "Rd", "Pu")) # specify 7 colors
+           box.palette = list("Gy", "Gn", "Bu", "Bn", 
+                              "Or", "Rd", "Pu")) 
 tree_full
 ```
 
@@ -339,7 +340,9 @@ fit <- Zoo |>
     method = "rpart",
     control = rpart.control(minsplit = 2), # we have little data
     tuneGrid = data.frame(cp = c(0.01, 0)),
-    trControl = trainControl(method = "LGOCV", p = 0.8, number = 10),
[…]
