+++ book/reference-keys.txt
@@ -39,7 +39,7 @@ jaccard-index
 distances-for-mixed-data
 gowers-coefficient
 using-euclidean-distance-with-mixed-data
-additional-proximity-measures-available-in-package-proxy
+more-proximity-measures
 data-exploration
 basic-statistics
 grouped-operations-and-calculations
@@ -61,7 +61,7 @@ correlation-matrix
 parallel-coordinates-plot
 more-visualizations
 exercises
-classification-basic-concepts-and-techniques
+classification-basic-concepts
 basic-concepts
 general-framework-for-classification
 the-zoo-dataset
@@ -110,21 +110,23 @@ iris-dataset
 nearest-neighbor-classifier-1
 naive-bayes-classifier-1
 linear-discriminant-analysis
-multinomial-logistic-regression-implemented-in-nnet
+multinomial-logistic-regression
 decision-trees
-svm
+ensemble-random-forest
+support-vector-machine
 single-layer-feed-forward-neural-networks
 circle-dataset
 nearest-neighbor-classifier-2
 naive-bayes-classifier-2
 linear-discriminant-analysis-1
-logistic-regression-implemented-in-nnet
+multinomial-logistic-regression-1
 decision-trees-1
-svm-1
+ensemble-random-forest-1
+support-vector-machine-1
 single-layer-feed-forward-neural-networks-1
 more-information-on-classification-with-r
 exercises-2
-association-analysis-basic-concepts-and-algorithms
+association-analysis-basic-concepts
 preliminaries
 the-arules-package
 transactions
@@ -132,7 +134,7 @@ create-transactions
 inspect-transactions
 vertical-layout-transaction-id-lists
 frequent-itemset-generation
-rules-generation
+rule-generation
 calculate-additional-interest-measures
 mine-using-templates
 compact-representation-of-frequent-itemsets
