+++ book_src/format_pkgs.R
@@ -3,5 +3,7 @@ if (!exists("all_pkgs"))
   all_pkgs <- character()
 
 format_pkgs <- function(pkgs)
-  paste(sapply(sort(pkgs), FUN = function(p) sprintf('_%s_ [@R-%s]', p ,p)), collapse = ', ')
+  paste(
+    sapply(sort(pkgs), FUN = function(p) sprintf('* _%s_ [@R-%s]', p, p)), 
+        collapse = '\n')
 
