+++ book_src/packages.bib
@@ -1,8 +1,8 @@
 @Manual{R-arules,
   title = {arules: Mining Association Rules and Frequent Itemsets},
   author = {Michael Hahsler and Christian Buchta and Bettina Gruen and Kurt Hornik},
-  year = {2023},
-  note = {R package version 1.7-7},
+  year = {2024},
+  note = {R package version 1.7-8},
   url = {https://github.com/mhahsler/arules},
 }
 
@@ -368,7 +368,7 @@ Regression},
   title = {seriation: Infrastructure for Ordering Objects Using Seriation},
   author = {Michael Hahsler and Christian Buchta and Kurt Hornik},
   year = {2024},
-  note = {R package version 1.5.5},
+  note = {R package version 1.5.6},
   url = {https://github.com/mhahsler/seriation},
 }
 
