### 2024-08-23 a61b4de
Use consistently CC-BY-SA.
