+++ README.md
@@ -10,10 +10,11 @@
 
 R and [tidyverse](https://www.tidyverse.org/) are very popular for data mining.
 This repository contains slides and documented R examples to accompany several
-chapters of the popular data mining text book:
+chapters of the popular data mining textbook:
 
 > Pang-Ning Tan, Michael Steinbach, Anuj Karpatne and Vipin Kumar, 
-[_Introduction to Data Mining,_](https://www-users.cs.umn.edu/~kumar001/dmbook/index.php) Addison Wesley, 1st or 2nd edition.
+> [_Introduction to Data Mining,_](https://www-users.cs.umn.edu/~kumar001/dmbook/index.php) 
+> Addison Wesley, 1st or 2nd edition.
 
 The slides and examples are used in my course _CS 5/7331 Data Mining_ taught at
 [SMU](https://www.smu.edu/) and will be regularly updated and improved.  The
@@ -24,24 +25,24 @@ which is published under the Creative Commons Attribution-ShareAlike license and
 share and adapt them freely. Please open an [issue](issues)  for corrections or
 to suggest improvements. 
 
-* [Browse the complete online book](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/)
+
+## Content
+
+* [Browse the online book](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/)
 * [Download the book as a PDF](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/R-Companion-Data-Mining.pdf)
 
-## Covered Chapters
-
-| Textbook Chapter* | Slides | Companion R Code |  
+| Companion Chapter | Lecture Slides | Free Textbook Chapter  |  
 | :--------| :---: | :----: |
-| 1. Introduction | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap1_intro.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/introduction.html) |
-| 2. Data | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html) | 
-| -. Web Chapter: [Exploring Data](https://www-users.cse.umn.edu/~kumar001/dmbook/data_exploration_1st_edition.pdf) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_exploration.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html#data-exploration)  | 
-| 3. [Classification: Basic Concepts and Techniques](https://www-users.cs.umn.edu/~kumar001/dmbook/ch3_classification.pdf) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap3_basic_classification.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-basic-concepts-and-techniques.html) | 
-| 4. Classification: Alternative Techniques |    [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap4_alternative_classification.pdf) | [R Code](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-alternative-techniques.html) |
[…]
