### 2024-08-28 #223 @stats-tgeorge
@openjournals/jose-eics I believe we are ready to move forward to publish. TY!
