+++ README.md
@@ -33,14 +33,14 @@ to suggest improvements.
 
 | Companion Chapter | Lecture Slides | Free Textbook Chapter  |  
 | :--------| :---: | :----: |
-| 1. [Introduction](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/introduction.html) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap1_intro.pdf) | -  |
-| 2. [Data](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pdf) | - | 
-| 2.5. [Exploring Data](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html#data-exploration)  | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_exploration.pdf) | Web Chapter [Exploring Data](https://www-users.cse.umn.edu/~kumar001/dmbook/data_exploration_1st_edition.pdf)  | 
-| 3. [Classification: Basic Concepts](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-basic-concepts.html) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap3_basic_classification.pdf) | [3. Classification](https://www-users.cs.umn.edu/~kumar001/dmbook/ch3_classification.pdf) | 
-| 4. [Classification: Alternative Techniques](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/classification-alternative-techniques.html) | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap4_alternative_classification.pdf) | - |
-| 5. [Association Analysis: Basic Concepts](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/association-analysis-basic-concepts.html)  | [Slides](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap5_basic_association_analysis.pdf) | [5. Association Analysis](https://www-users.cs.umn.edu/~kumar001/dmbook/ch5_association_analysis.pdf) |
+| 1. [Introduction](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/introduction.html) | [PDF](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap1_intro.pdf), [PowerPoint](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap1_intro.pptx) | -  |
+| 2. [Data](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/data.html) | [PDF](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pdf), [PowerPoint](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/slides/chap2_data.pptx) | - | 
[…]
