+++ book_src/02_data.Rmd
@@ -46,7 +46,7 @@ simple labels all the way to numbers. The scales come in four levels.
 | Scale Name | Description | Operations | Statistics | R |
 | ---------- | ------------| --------- | -------| ------- |
 | Nominal | just a label (e.g., red, green) | $==, !=$ | counts | `factor` |
-| Ordinal | label with order (e.g., small, med., large) | $<, >$ | median | `ordered` | 
+| Ordinal | label with order (e.g., small, med., large) | $<, >$ | median | `ordered factor` | 
 | Interval | difference between two values is meaningful (regular number) | $+, -$ | mean, sd | `numeric` |
 | Ratio | has a natural zero (e.g., count, distance) | $/, *$ | percent | `numeric` |
 
@@ -54,17 +54,26 @@ The scales build on each other meaning that an ordinal variable also has the cha
 a nominal variable with the added order information.
 We often do not differentiate between interval and ratio scale because we rarely 
 not need to calculate percentages or other statistics that require a meaningful zero value.
-The following code gives example vectors with nominal, ordinal (`levels` defines the order) 
-and interval scale.  
+
+Nominal data is created using `factor()`. If the factor levels
+are not specified, then they are created in alphabetical order.
 
 ```{r}
 factor(c("red", "green", "green", "blue"))
+```
 
-factor("S", "L", "M", "S", 
-       levels = c("S", "M", "L"), 
-       ordered = TRUE)
+Ordinal data is created using `ordered()`. The levels specify the order. 
 
-c(1, 2, 3, 4, 3)
+```{r}
+ordered(c("S", "L", "M", "S"), 
+       levels = c("S", "M", "L"))
+```
+
+
+Ratio/interval data is created as a simple vector.
+
+```{r}
+c(1, 2, 3, 4, 3, 3)
 ```
 
 
