### 2024-09-25 5d4e6c7
Added option for histogramm and star plots.
