+++ book_src/packages.bib
@@ -8,9 +8,9 @@
 
 @Manual{R-arulesSequences,
   title = {arulesSequences: Mining Frequent Sequences},
-  author = {Christian Buchta and Michael Hahsler and with contributions from Daniel Diaz},
-  year = {2023},
-  note = {R package version 0.2-30},
+  author = {Christian Buchta and Michael Hahsler},
+  year = {2024},
+  note = {R package version 0.2-31},
   url = {https://CRAN.R-project.org/package=arulesSequences},
 }
 
@@ -85,8 +85,8 @@ al.},
   title = {e1071: Misc Functions of the Department of Statistics, Probability
 Theory Group (Formerly: E1071), TU Wien},
   author = {David Meyer and Evgenia Dimitriadou and Kurt Hornik and Andreas Weingessel and Friedrich Leisch},
-  year = {2023},
-  note = {R package version 1.7-14},
+  year = {2024},
+  note = {R package version 1.7-16},
   url = {https://CRAN.R-project.org/package=e1071},
 }
 
@@ -110,7 +110,7 @@ Theory Group (Formerly: E1071), TU Wien},
   title = {fpc: Flexible Procedures for Clustering},
   author = {Christian Hennig},
   year = {2024},
-  note = {R package version 2.2-12},
+  note = {R package version 2.2-13},
   url = {https://www.unibo.it/sitoweb/christian.hennig/en/},
 }
 
@@ -157,24 +157,24 @@ Theory Group (Formerly: E1071), TU Wien},
 @Manual{R-hexbin,
   title = {hexbin: Hexagonal Binning Routines},
   author = {Dan Carr and Nicholas Lewin-Koh and Martin Maechler},
-  year = {2023},
-  note = {R package version 1.28.3},
+  year = {2024},
+  note = {R package version 1.28.4},
   url = {https://github.com/edzer/hexbin},
 }
 
-@Manual{R-keras,
-  title = {keras: R Interface to Keras},
-  author = {JJ Allaire and François Chollet},
+@Manual{R-keras3,
+  title = {keras3: R Interface to Keras},
+  author = {Tomasz Kalinowski and JJ Allaire and François Chollet},
   year = {2024},
-  note = {R package version 2.15.0},
-  url = {https://tensorflow.rstudio.com/},
+  note = {R package version 1.2.0},
+  url = {https://keras3.posit.co/},
 }
 
 @Manual{R-kernlab,
   title = {kernlab: Kernel-Based Machine Learning Lab},
   author = {Alexandros Karatzoglou and Alex Smola and Kurt Hornik},
-  year = {2023},
-  note = {R package version 0.9-32},
+  year = {2024},
+  note = {R package version 0.9-33},
   url = {https://CRAN.R-project.org/package=kernlab},
 }
 
@@ -222,7 +222,7 @@ Theory Group (Formerly: E1071), TU Wien},
   title = {Matrix: Sparse and Dense Matrix Classes and Methods},
   author = {Douglas Bates and Martin Maechler and Mikael Jagan},
   year = {2024},
-  note = {R package version 1.6-5},
+  note = {R package version 1.7-0},
   url = {https://Matrix.R-forge.R-project.org},
 }
 
@@ -263,7 +263,7 @@ Classification, and Density Estimation},
   title = {party: A Laboratory for Recursive Partytioning},
   author = {Torsten Hothorn and Kurt Hornik and Carolin Strobl and Achim Zeileis},
   year = {2024},
-  note = {R package version 1.3-15},
+  note = {R package version 1.3-17},
   url = {http://party.R-forge.R-project.org},
 }
 
[…]
