### 2024-09-30 086d63b
Added ggplot2 intro to chapter 2.
