+++ book_src/packages.bib
@@ -162,14 +162,6 @@ Theory Group (Formerly: E1071), TU Wien},
   url = {https://github.com/edzer/hexbin},
 }
 
-@Manual{R-keras3,
-  title = {keras3: R Interface to Keras},
-  author = {Tomasz Kalinowski and JJ Allaire and François Chollet},
-  year = {2024},
-  note = {R package version 1.2.0},
-  url = {https://keras3.posit.co/},
-}
-
 @Manual{R-kernlab,
   title = {kernlab: Kernel-Based Machine Learning Lab},
   author = {Alexandros Karatzoglou and Alex Smola and Kurt Hornik},
