### 2024-10-09 854d5b5
Show the k-means object.
