+++ book_src/07_cluster_analysis.Rmd
@@ -224,42 +224,57 @@ fviz_cluster(kmeans(ruspini_scaled, centers = 8), data = ruspini_scaled,
 ## Agglomerative Hierarchical Clustering
 
 Hierarchical clustering starts with a distance matrix. `dist()` defaults
-to method="Euclidean". **Note:** Distance matrices become very large
-quickly (size and time complexity is $O(n^2)$ where $n$ is the number if
+to `method = "euclidean"`. 
+
+**Notes:** 
+
+* Distance matrices become very large
+quickly (the space complexity is $O(n^2)$ where $n$ is the number if
 data points). It is only possible to calculate and store the matrix for
-small data sets (maybe a few hundred thousand data points) in main
-memory. If your data is too large then you can use sampling.
+small to medium data sets (maybe a few hundred thousand data points) in main
+memory. If your data is too large then you can use sampling to reduce the 
+number of points to cluster.
+* The data needs to be scaled since we compute distances.
+
+
+### Creating a Dendrogram
 
 ```{r }
 d <- dist(ruspini_scaled)
 ```
 
 `hclust()` implements [agglomerative hierarchical
-clustering](https://en.wikipedia.org/wiki/Hierarchical_clustering). We
-cluster using complete link.
+clustering](https://en.wikipedia.org/wiki/Hierarchical_clustering). 
+The linkage function defines how the distance between two clusters (sets of points) 
+is calculated. We specify the complete-link method where the distance between two groups 
+is measured by the distance between the two farthest apart points in the two groups.
 
 ```{r }
 hc <- hclust(d, method = "complete")
 ```
 
 Hierarchical clustering does not return cluster assignments but a
-dendrogram. The standard plot function plots the dendrogram.
+dendrogram object which shows how the objects on the x-axis are successively joined together
+when going up along the y-axis. The y-axis represents the distance at which groups of points
+are joined together. The standard plot function displays the dendrogram.
 
 ```{r }
 plot(hc)
 ```
 
-Use `factoextra` (ggplot version). We can specify the number of clusters
-to visualize how the dendrogram will be cut into clusters.
+The package `factoextra` provides a ggplot function to plot dendrograms. 
 
 ```{r warning=FALSE}
-fviz_dend(hc, k = 4)
+fviz_dend(hc)
 ```
 
 More plotting options for dendrograms, including plotting parts of large
 dendrograms can be found [here.](https://rpubs.com/gaston/dendrograms)
 
-Extract cluster assignments by cutting the dendrogram into four parts
+### Extracting a Partitional Clustering
+
+Partitional clusterings with cluster assignments can be extracted from a dendrogram
+by cutting the dendrogram horizontally using `cutree()`. Here we cut it into four parts
 and add the cluster id to the data.
 
 ```{r }
@@ -267,21 +282,33 @@ clusters <- cutree(hc, k = 4)
 cluster_complete <- ruspini_scaled |>
   add_column(cluster = factor(clusters))
 cluster_complete
+```
 
+The 4 different clusters can be shown in the dendrogram using different colors.  
+
+```{r warning=FALSE}
+fviz_dend(hc, k = 4)
+```
+
+We can also color in the scatterplot.
+
+```{r}
 ggplot(cluster_complete, aes(x, y, color = cluster)) +
   geom_point()
 ```
 
-Try 8 clusters (Note: `fviz_cluster` needs a list with data and the
-cluster labels for hclust)
+The package `factoextra` provides an alternative visualization. Since it needs
+a partition object as the first argument, we create one as a simple list.
 
 ```{r }
 fviz_cluster(list(data = ruspini_scaled, 
-                  cluster = cutree(hc, k = 8)), 
+                  cluster = cutree(hc, k = 4)), 
[…]
