### 2024-10-15 bc6b510
updated clustering chapter.
