+++ book/reference-keys.txt
@@ -161,7 +161,6 @@ data-cleaning
 scale-data
 k-means
 inspect-clusters
-cluster-profiles
 extract-a-single-cluster
 agglomerative-hierarchical-clustering
 creating-a-dendrogram
