+++ book_src/07_cluster_analysis.Rmd
@@ -422,22 +422,18 @@ squares (WCSS) used as the optimization objective by
 width](https://en.wikipedia.org/wiki/Silhouette_(clustering)). Look at
 `within.cluster.ss` and `avg.silwidth` below.
 
-```{r }
-##library(fpc)
-```
+Some notes about the code: 
 
-Notes: 
-
-* I do not load fpc since the NAMESPACE overwrites dbscan. 
+* I do not load the package `fpc` using `library()` since it would mask the `dbscan()` function in package `dbscan`. Instead I use the namespace operator `::`. 
 * The clustering (second argument below) has to be supplied as a vector
-with numbers (cluster IDs) and cannot be a factor (use `as.integer()` to
-convert the factor to an ID).
+of integers (cluster IDs) and cannot be a factor (to make sure, we can use `as.integer()`).
 
 ```{r }
-fpc::cluster.stats(d, km$cluster)
+# library(fpc)
+fpc::cluster.stats(d, as.integer(km$cluster))
 ```
 
-Read `? cluster.stats` for an explanation of all the available indices.
+Read the man page for `cluster.stats()` for an explanation of all the available indices.
 
 ```{r }
 sapply(
@@ -489,7 +485,7 @@ the color representing the value.
 
 ```{r }
 library(seriation)
-pimage(d, col = bluered(100))
+ggpimage(d)
 ```
 
 Rows and columns are the objects as they are ordered in the data set.
@@ -501,7 +497,7 @@ matrix (rows and columns) using the k-means cluster labels from cluster
 visible.
 
 ```{r }
-pimage(d, order=order(km$cluster), col = bluered(100))
+ggpimage(d, order=order(km$cluster))
 ```
 
 Plot function `dissplot` in package **seriation** rearranges the matrix
@@ -511,7 +507,7 @@ the objects by cluster and then reorders clusters and objects within
 clusters so that more similar objects are closer together.
 
 ```{r }
-dissplot(d, labels = km$cluster, 
+ggdissplot(d, labels = km$cluster, 
          options = list(main = "k-means with k=4"))
 ```
 
@@ -519,13 +515,11 @@ The reordering by `dissplot` makes the misspecification of k visible as
 blocks.
 
 ```{r }
-dissplot(d, 
-         labels = kmeans(ruspini_scaled, centers = 3)$cluster, 
-         col = bluered(100))
+ggdissplot(d, 
+         labels = kmeans(ruspini_scaled, centers = 3)$cluster)
 
-dissplot(d, 
-         labels = kmeans(ruspini_scaled, centers = 9)$cluster, 
-         col = bluered(100))
+ggdissplot(d, 
+         labels = kmeans(ruspini_scaled, centers = 9)$cluster)
 ```
 
 Using `factoextra`
@@ -669,7 +663,7 @@ before using Euclidean distance.
 library(seriation)
 
 d_shapes <- dist(scale(shapes))
-VAT(d_shapes, col = bluered(100))
+ggVAT(d_shapes)
 ```
 
 iVAT uses the largest distances for all possible paths between two
@@ -677,7 +671,7 @@ objects instead of the direct distances to make the block structure
 better visible.
 
 ```{r }
-iVAT(d_shapes, col = bluered(100))
+ggiVAT(d_shapes)
 ```
 
 #### Hopkins statistic
@@ -708,8 +702,8 @@ No point clouds are visible, just noise.
 
 ```{r }
 d_random <- dist(data_random)
-VAT(d_random, col = bluered(100))
-iVAT(d_random, col = bluered(100))
+ggVAT(d_random)
+ggiVAT(d_random)
[…]
