### 2024-10-25 8e653cc
rewrote purity and entropy function with comments.
