+++ book_src/07_cluster_analysis.Rmd
@@ -843,34 +843,41 @@ and
 [purity](https://en.wikipedia.org/wiki/Cluster_analysis#External_evaluation).
 
 `cluster_stats()` computes ARI and VI as comparative measures. I define
-functions for entropy and purity here:
+functions for the purity and the entropy of a clustering given the ground truth here:
 
 ```{r }
-entropy <- function(cluster, truth) {
-  k <- max(cluster, truth)
-  cluster <- factor(cluster, levels = 1:k)
-  truth <- factor(truth, levels = 1:k)
-  w <- table(cluster)/length(cluster)
-
-  cnts <- sapply(split(truth, cluster), table)
-  p <- sweep(cnts, 1, rowSums(cnts), "/")
-  p[is.nan(p)] <- 0
-  e <- -p * log(p, 2)
-
-  sum(w * rowSums(e, na.rm = TRUE))
+purity <- function(cluster, truth, show_table = FALSE) {
+  if (length(cluster) != length(truth))
+    stop("Cluster vector and ground truth vectors are not of the same length!")
+  
+  # tabulate
+  tbl <- table(cluster, truth)
+  if(show_table)
+    print(tbl)
+  
+  # find majority class
+  majority <- apply(tbl, 1, max)
+  sum(majority) / length(cluster)
 }
 
-purity <- function(cluster, truth) {
-  k <- max(cluster, truth)
-  cluster <- factor(cluster, levels = 1:k)
-  truth <- factor(truth, levels = 1:k)
-  w <- table(cluster)/length(cluster)
+entropy <- function(cluster, truth, show_table = FALSE) {
+  if (length(cluster) != length(truth))
+    stop("Cluster vector and ground truth vectors are not of the same length!")
+  
+  # calculate membership probability of cluster to class
+  tbl <- table(cluster, truth)
+  p <- sweep(tbl, 2, colSums(tbl), "/")
+  
+  if(show_table)
+    print(p)
 
-  cnts <- sapply(split(truth, cluster), table)
-  p <- sweep(cnts, 1, rowSums(cnts), "/")
-  p[is.nan(p)] <- 0
-
-  sum(w * apply(p, 1, max))
+  # calculate cluster entropy
+  e <- -p * log(p, 2)
+  e <- rowSums(e, na.rm = TRUE)
+  
+  # weighted sum over clusters
+  w <- table(cluster) / length(cluster)
+  sum(w * e)
 }
 ```
 
@@ -882,29 +889,36 @@ random_4 <- sample(1:4, nrow(shapes), replace = TRUE)
 random_6 <- sample(1:6, nrow(shapes), replace = TRUE)
 
 r <- rbind(
+  truth = c(
+    unlist(fpc::cluster.stats(d, truth, 
+                              truth, compareonly = TRUE)),
+    purity = purity(truth, truth),
+    entropy = entropy(truth, truth)
+  ),
+  
   kmeans_7 = c(
     unlist(fpc::cluster.stats(d, km$cluster, 
                               truth, compareonly = TRUE)),
-    entropy = entropy(km$cluster, truth),
-    purity = purity(km$cluster, truth)
+    purity = purity(km$cluster, truth),
+    entropy = entropy(km$cluster, truth)
     ),
   hc_4 = c(
     unlist(fpc::cluster.stats(d, hc_4, 
                               truth, compareonly = TRUE)),
-    entropy = entropy(hc_4, truth),
-    purity = purity(hc_4, truth)
+    purity = purity(hc_4, truth),
+    entropy = entropy(hc_4, truth)
     ),
   random_4 = c(
     unlist(fpc::cluster.stats(d, random_4, 
                               truth, compareonly = TRUE)),
[…]
