### 2024-10-25 #223 @mhahsler
@stats-tgeorge  Hi George. Is the @openjournals/jose-eics handle correct?
