### 2024-11-11 #223 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
