### 2024-11-11 #223 @stats-tgeorge
> @stats-tgeorge Hi George. Is the @openjournals/jose-eics handle correct?

You were correct. I was following our guide (which suggested what I did) and not my checklist. My mistake.
