### 2024-11-11 #223 @labarba
Should the archive DOI be updated to `v2`?
