### 2024-11-11 #223 @stats-tgeorge
It looks like version 2 was created after the DOI was updated here. Yes, it now needs to be updated again.
