### 2024-11-11 #223 @labarba
I'm confused. This archive:

> Hahsler, Michael (2024). An R Companion for Introduction to Data Mining. figshare. Book. https://doi.org/10.6084/m9.figshare.26750404.v2

...contains a 357-page PDF.

`v1` of that archive seems to hold the file contents of a website.

In the JOSE paper, it says that the materials are available at https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/

The "View book source" button sends me to this GitHub repo: https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples

So why don't we have as the archive a deposit of the GitHub repository, which is in fact the _source_ for the book? If a user wants to fork and modify, they need the source.

I also notice that the Figshare deposit (both versions) show a CC-BY license, while the website and GitHub repo indicate CC-BY-SA. Shouldn't the licenses be the same? (I realize these are different objects, per my inquiry above, but still…)
