### 2024-11-12 #223 @stats-tgeorge
@labarba should @mhahsler then create a V3 of the archive that has the appropriate items (archive for the GitHub repo)? Then pick a singular license?
