### 2024-11-12 8230bb9
small improvements and changes license.
