+++ book_src/packages.bib
@@ -214,7 +214,7 @@ Theory Group (Formerly: E1071), TU Wien},
   title = {Matrix: Sparse and Dense Matrix Classes and Methods},
   author = {Douglas Bates and Martin Maechler and Mikael Jagan},
   year = {2024},
-  note = {R package version 1.7-0},
+  note = {R package version 1.7-1},
   url = {https://Matrix.R-forge.R-project.org},
 }
 
