### 2024-11-12 #223 @labarba
I notice the [comment above](https://github.com/openjournals/jose-reviews/issues/223#issuecomment-2291898680) where the author says:

> Note on license: figshare only gives a choice for CC BY 4.0, while the book uses CC BY-NC 4.0. I can change everything to CC BY 4.0, if that is necessary.

The [online book](https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/) shows a CC-BY license, so this matches. I had seen on the repository a CC-BY-SA license, but the author has just [changed](https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples/commit/8230bb92fef748ca8a6c54112795be01f76e856d) it. Indeed, the license should match everywhere.

The issue remains that the archive on Figshare is a collection of website files (v1) or a PDF (v2). What JOSE wants is an archive of the _source_, which is the only thing that guarantees future reuse and derivative works. Having a Fighsare deposit of the PDF is nice (users can immediately read), but it is not conducive to derivative works.

I would suggest a Zenodo deposit of the GitHub repo. If the author prefers Figshare, then a zip of the repo is the only way.
