### 2024-11-12 #223 @mhahsler
@labarba  and @stats-tgeorge: I agree this isn't very clear.

I have now done the following:
1. Figshare does not give the option of using CC-BY-SA only CC-BY, so I have changed the license for all the material to the more permissive CS-BY license. This needs to be updated in the paper.
2. I have made a new version on Figshare (v3) which contains the books PDF and a archived version of the complete GitHub repository + a README with a link to GitHub. The paper should only use https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples and  https://doi.org/10.6084/m9.figshare.26750404.v3

@labarba Please let me know if I should update the paper in my GitHub repository.

BTW: I tried Zenodo first, but it did not like the repository with so many non-code files and died in the process of importing the repository.

Thanks for your help,
Michael
