### 2024-11-18 2a04763
Improved wording for classification and regression.
