+++ book_src/03_classification_basics.Rmd
@@ -272,9 +272,9 @@ library(caret)
 confusionMatrix(data = pred, 
                 reference = Zoo |> pull(type))
 ```
-**Note:** Calculating accuracy on the training data is not a good idea. 
-Keep on reading!
-
+**Important note:** Calculating accuracy on the training data is not a good idea. 
+A complete example with code for holding out a test set and performing hyperparameter
+selection using cross-validation can be found in section [Hyperparameter Tuning].
 
 ## Model Overfitting
 
@@ -414,6 +414,10 @@ parallel. To enable multi-core support, `caret` uses the package
 
 ## Hyperparameter Tuning
 
+**Note:** This section contains a complete code example of how data should be used.
+It first holds out a test set and then performing hyperparameter selection 
+using cross-validation.
+
 Hyperparameters are parameters that change how a
 training algorithm works. An example is the complexity parameter 
 `cp` for rpart decision trees. Tuning the hyperparameter means that 
@@ -541,8 +545,9 @@ confusionMatrix(data = pred,
 
 ## Model Comparison
 
-We will compare decision trees with a k-nearest neighbors (kNN)
-classifier. We will create fixed sampling scheme (10-folds) so we
+We will compare three models, a majority class baseline classifier, a 
+decision trees with a k-nearest neighbors (kNN)
+classifier. We will create fixed sampling scheme (10-fold cross-validation) so we
 compare the different models using exactly the same folds. It is
 specified as `trControl` during training.
 
@@ -552,7 +557,50 @@ train_index <- createFolds(Zoo_train$type,
                            returnTrain = TRUE)
 ```
 
-Build models
+### Build models
+
+Caret does not provide a baseline classifier, but lets us specify our own models. 
+We define the classifier as a list with the  
+structure needed by Caret's `train()` function (see [use your own model](https://topepo.github.io/caret/using-your-own-model-in-train.html#model-components)  in caret's description). The `fit` component calculates
+the majority class label and the `predict` component just labels every row
+in `newdata` as the majority class.
+
+```{r }
+majority_class_baseline_classifier <- list(label = "Majority Class Baseline Classifier",
+                  library = character(0),
+                  type = "Classification",
+                  parameters = data.frame(parameter = "parameter",
+                                          class = "character",
+                                          label = "parameter"),
+                  grid = function(x, y, len = NULL, search = "grid") 
+                    data.frame(parameter = "none"),
+                  loop = NULL,
+                  fit = function(x, y, wts, param, lev, last, classProbs, ...) {          
+                    tbl <- table(y)
+                    majority <- names(tbl)[which.max(tbl)]
+                    list(majority = majority)
+                    },
+                  predict = function(modelFit, newdata, submodels = NULL) {
+                    rep(modelFit$majority, nrow(newdata))
+                  },
+                  prob = NULL
+                  )
+```  
+  
+We train the model and perform cross-validation.
+```{r}  
+baseline <-  Zoo_train |> 
+  train(type ~ .,
+        data = _,
+        method = majority_class_baseline_classifier,
+	      tuneLength = 1,
+	      trControl = trainControl(method = "cv", 
+	                               index = train_index)
+  )
+```
+
+
+The second model is a default decision tree.
 
 ```{r }
 rpartFit <- Zoo_train |> 
@@ -560,14 +608,17 @@ rpartFit <- Zoo_train |>
         data = _,
         method = "rpart",
         tuneLength = 10,
-        trControl = trainControl(method = "cv", 
+        trControl = trainControl(method = "cv",
                                  index = train_index)
   )
 ```
 
[…]
