+++ book/reference-keys.txt
@@ -1,0 +1,214 @@
+introduction
+used-software
+base-r
+vectors
+vectorized-operations
+subsetting-vectors
+functions
+strings
+plotting
+objects
+r-markdown
+tidyverse
+tibbles
+transformations
+ggplot2
+data
+types-of-data
+attributes-and-measurement
+the-iris-dataset
+data-quality
+data-preprocessing
+aggregation
+sampling
+random-sampling
+stratified-sampling
+dimensionality-reduction
+principal-components-analysis-pca
+multi-dimensional-scaling-mds
+non-parametric-multidimensional-scaling
+other-nonlinear-dimensionality-reduction-methods
+variable-transformation-standardization
+measures-of-similarity-and-dissimilarity
+minkowsky-distances
+distances-for-binary-data
+hamming-distance
+jaccard-index
+distances-for-mixed-data
+gowers-coefficient
+using-euclidean-distance-with-mixed-data
+more-proximity-measures
+data-exploration
+basic-statistics
+grouped-operations-and-calculations
+tabulate-data
+percentiles-quantiles
+correlation
+pearson-correlation
+rank-correlation
+density
+histograms
+kernel-density-estimate-kde
+visualization
+histogram
+boxplot
+scatter-plot
+scatter-plot-matrix
+matrix-visualization
+correlation-matrix
+parallel-coordinates-plot
+star-plot
+more-visualizations
+exercises
+classification-basic-concepts
+basic-concepts
+general-framework-for-classification
+the-zoo-dataset
+decision-tree-classifiers
+create-tree
+make-predictions-for-new-data
+manual-calculation-of-the-resubstitution-error
+confusion-matrix-for-the-resubstitution-error-using-caret
+model-overfitting
+model-selection
+model-evaluation
+holdout-method
+cross-validation-methods
+hyperparameter-tuning
+pitfalls-of-model-selection-and-evaluation
+model-comparison
+build-models
+feature-selection
+univariate-feature-importance-score
+feature-subset-selection
+using-dummy-variables-for-factors
+exercises-1
+classification-alternative-techniques
+types-of-classifiers
+set-up-the-training-and-test-data
+rule-based-classifier-part
+nearest-neighbor-classifier
+naive-bayes-classifier
+bayesian-network
+logistic-regression
+artificial-neural-network-ann
+support-vector-machines
+ensemble-methods
+random-forest
+gradient-boosted-decision-trees-xgboost
+class-imbalance
+option-1-use-the-data-as-is-and-hope-for-the-best
+option-2-balance-data-with-resampling
+option-3-build-a-larger-tree-and-use-predicted-probabilities
+create-a-biased-classifier
+plot-the-roc-curve
+option-4-use-a-cost-sensitive-classifier
+model-comparison-1
+comparing-decision-boundaries-of-popular-classification-techniques
+iris-dataset
+nearest-neighbor-classifier-1
+naive-bayes-classifier-1
+linear-discriminant-analysis
+multinomial-logistic-regression
+decision-trees
+ensemble-random-forest
+support-vector-machine
+single-layer-feed-forward-neural-networks
+circle-dataset
+nearest-neighbor-classifier-2
+naive-bayes-classifier-2
+linear-discriminant-analysis-1
+multinomial-logistic-regression-1
+decision-trees-1
+ensemble-random-forest-1
+support-vector-machine-1
+single-layer-feed-forward-neural-networks-1
+more-information-on-classification-with-r
+exercises-2
+association-analysis-basic-concepts
+preliminaries
+the-arules-package
+transactions
+create-transactions
+inspect-transactions
+vertical-layout-transaction-id-lists
+frequent-itemset-generation
+rule-generation
[…]
