### 2024-11-22 11dfea1
improved model comparison.
