+++ book_src/04_classification_alternative.Rmd
@@ -91,23 +91,7 @@ Zoo_train <- Zoo |> slice(inTrain)
 Zoo_test <- Zoo |> slice(-inTrain)
 ```
 
-`train()` is aware of the tunable hyperparameters of each method and 
-automatically performs model selection using a validation set. 
-We will use for all models the same validation sets by
-creating a fixed sampling scheme (10-folds). This will help when we can compare the fitted models later.
-
-```{r }
-train_index <- createFolds(Zoo_train$type, 
-                           k = 10, 
-                           returnTrain = TRUE)
-```
-
-The fixed folds are used in `train()` with the argument
-`trControl = trainControl(method = "cv", index = train_index))`. If
-you don't need fixed folds, then remove `index = train_index` in the
-code below.
-
-For help with building models in caret see: `? train`
+For hyperparameter tuning, we will use 10-fold cross-validation.
 
 **Note:** Be careful if you have many `NA` values in your data.
 `train()` and cross-validation many fail in some cases. If that is the
@@ -125,8 +109,7 @@ rulesFit <- Zoo_train |> train(type ~ .,
   method = "PART",
   data = _,
   tuneLength = 5,
-  trControl = trainControl(method = "cv", 
-                           index = train_index))
+  trControl = trainControl(method = "cv"))
 rulesFit
 ```
 The model selection results are shown in the table. This is the selected model:
@@ -157,8 +140,7 @@ knnFit <- Zoo_train |> train(type ~ .,
   data = _,
   preProcess = "scale",
   tuneGrid = data.frame(k = c(1, 3, 5, 7, 9)),
-	trControl = trainControl(method = "cv", 
-	                         index = train_index))
+	trControl = trainControl(method = "cv"))
 knnFit
 ```
 
@@ -181,8 +163,7 @@ NBFit <- train(x = as.data.frame(Zoo_train[, -ncol(Zoo_train)]),
                method = "nb",
                tuneGrid = data.frame(fL = c(.2, .5, 1, 5), 
                                      usekernel = TRUE, adjust = 1),
-               trControl = trainControl(method = "cv", 
-                                        index = train_index))
+               trControl = trainControl(method = "cv"))
 NBFit
 ```
 
@@ -228,8 +209,7 @@ logRegFit <- Zoo_train |> train(type ~ .,
   method = "multinom",
   data = _,
   trace = FALSE, # suppress some output
-	trControl = trainControl(method = "cv", 
-	                         index = train_index))
+	trControl = trainControl(method = "cv"))
 logRegFit
 ```
 
@@ -252,8 +232,7 @@ nnetFit <- Zoo_train |> train(type ~ .,
   method = "nnet",
   data = _,
 	tuneLength = 5,
-	trControl = trainControl(method = "cv", 
-	                         index = train_index),
+	trControl = trainControl(method = "cv"),
   trace = FALSE # no progress output
   )
 nnetFit
@@ -282,8 +261,7 @@ svmFit <- Zoo_train |> train(type ~.,
   method = "svmLinear",
   data = _,
 	tuneLength = 5,
-	trControl = trainControl(method = "cv", 
-	                         index = train_index))
+	trControl = trainControl(method = "cv"))
 svmFit
 ```
 
@@ -309,8 +287,7 @@ randomForestFit <- Zoo_train |> train(type ~ .,
   method = "rf",
   data = _,
 	tuneLength = 5,
-	trControl = trainControl(method = "cv", 
-	                         index = train_index))
+	trControl = trainControl(method = "cv"))
 randomForestFit
 ```
 
[…]
