+++ book_src/03_classification_basics.Rmd
@@ -16,7 +16,7 @@ the data mining textbook is available online:
 ### Packages Used in this Chapter {.unnumbered}
 
 ```{r setup_03}
-pkgs <- c("caret", "FSelector", "lattice", "mlbench", 
+pkgs <- c("basemodels", "caret", "FSelector", "lattice", "mlbench", 
           "palmerpenguins", "party", "pROC", "rpart", 
           "rpart.plot", "tidyverse")
 
@@ -554,43 +554,19 @@ results can be used to compare models statistically.
 
 ### Build models
 
-Caret does not provide a baseline classifier, but lets us specify our own models. 
-We define the classifier as a list with the  
-structure needed by Caret's `train()` function (see [use your own model](https://topepo.github.io/caret/using-your-own-model-in-train.html#model-components)  in caret's description). The `fit` component calculates
-the majority class label and the `predict` component just labels every row
-in `newdata` as the majority class.
+Caret does not provide a baseline classifier, but the package `basemodels` does.
+We first create a weak baseline model that always predicts the the majority 
+class mammal.
 
 ```{r }
-majority_class_baseline_classifier <- list(label = "Majority Class Baseline Classifier",
-                  library = character(0),
-                  type = "Classification",
-                  parameters = data.frame(parameter = "parameter",
-                                          class = "character",
-                                          label = "parameter"),
-                  grid = function(x, y, len = NULL, search = "grid") 
-                    data.frame(parameter = "none"),
-                  loop = NULL,
-                  fit = function(x, y, wts, param, lev, last, classProbs, ...) {          
-                    tbl <- table(y)
-                    majority <- names(tbl)[which.max(tbl)]
-                    list(majority = majority)
-                    },
-                  predict = function(modelFit, newdata, submodels = NULL) {
-                    rep(modelFit$majority, nrow(newdata))
-                  },
-                  prob = NULL
-                  )
-```  
-  
-We train the model and perform cross-validation.
-```{r}  
-baseline <-  Zoo_train |> 
-  train(type ~ .,
-        data = _,
-        method = majority_class_baseline_classifier,
-	      tuneLength = 1, ### there are no hyperparameters
-	      trControl = trainControl(method = "cv")
-  )
+baseline <- Zoo_train |> train(type ~ .,
+  method = basemodels::dummyClassifier,
+  data = _,
+  strategy = "constant",
+  constant = "mammal",
+  trControl = trainControl(method = "cv" 
+                           ))
+baseline
 ```
 
 
@@ -604,6 +580,7 @@ rpartFit <- Zoo_train |>
         tuneLength = 10,
         trControl = trainControl(method = "cv")
   )
+rpartFit
 ```
 
 The third model is a kNN classifier, this classifier will be discussed 
@@ -622,6 +599,7 @@ knnFit <- Zoo_train |>
 	      tuneLength = 10,
 	      trControl = trainControl(method = "cv")
   )
+knnFit
 ```
 
 
