+++ book_src/packages.bib
@@ -359,9 +359,9 @@ Regression},
 @Manual{R-seriation,
   title = {seriation: Infrastructure for Ordering Objects Using Seriation},
   author = {Michael Hahsler and Christian Buchta and Kurt Hornik},
-  year = {2024},
-  note = {R package version 1.5.6},
+  note = {R package version 1.5.6.1},
   url = {https://github.com/mhahsler/seriation},
+  year = {2024},
 }
 
 @Manual{R-stringr,

[…]
