+++ book_src/packages.bib
@@ -1,9 +1,9 @@
 @Manual{R-arules,
   title = {arules: Mining Association Rules and Frequent Itemsets},
   author = {Michael Hahsler and Christian Buchta and Bettina Gruen and Kurt Hornik},
-  year = {2024},
-  note = {R package version 1.7-8},
+  note = {R package version 1.7-8-1},
   url = {https://github.com/mhahsler/arules},
+  year = {2024},
 }
 
 @Manual{R-arulesSequences,
