### 2024-12-24 2686939
update license and repo link
plus some small copy edits
