+++ paper/paper.md
@@ -26,13 +26,13 @@ association analysis.  The resource uses complete, annotated examples to
 demonstrate how data mining concepts can be translated into R code.
 
 
-The materials have been made publicly available at: <https://mhahsler.github.io/Introduction_to_Data_Mining_R_Examples/book/> and licensed under the [Creative Commons Attribution-NonCommercial-ShareAlike 4.0 (CC BY-NC-SA 4.0) License](https://creativecommons.org/licenses/by-nc-sa/4.0/).
+The materials have been made publicly available at: <https://github.com/mhahsler/Introduction_to_Data_Mining_R_Examples> and licensed under the [Creative Commons Attribution 4.0 (CC BY 4.0) License](https://creativecommons.org/licenses/by/4.0/).
 
 # Statement of Need
 
-The textbook Introduction to Data Mining [@Tan2018] has been one of the most
-popular choices to learn and teach data mining concepts.  Some of the most
-important chapters have been made available for free by the authors on the
+The textbook _Introduction to Data Mining_ [@Tan2018] has been one of the most
+popular choices to learn and teach data mining concepts.  Several
+chapters have been made available for free by the authors on the
 [books's website](https://www-users.cse.umn.edu/~kumar001/dmbook/index.php).
 One of the authors also provides Python Jupyter notebooks with examples, but
 complete R code examples were still needed. Given the R community's interest in
@@ -45,7 +45,7 @@ mining.
 # Learning Objectives and Content
 
 The resource assumes basic knowledge of programming and statistics.  The
-learning objectives are to learn how to
+learning objectives are to:
 
 * prepare and understand data,
 * perform classification,
@@ -62,10 +62,10 @@ preparing students to work with more complex real-world datasets by initially
 relieving them from dealing with too many low-level implementation details
 while exploring the concepts. 
 
-The resource mirrors the textbook's structure so it can be easily used along
-with the textbook. After a short introduction, Chapter 2 discusses data types
-in R, data quality concerns, and data preprocessing. In addition, data
-exploration and visualization examples are given. Chapters 3 and 4 cover
+The resource mirrors the textbook's structure so it can be used along
+with it easily. After a short introduction, Chapter 2 discusses data types
+in R, data quality concerns, and data preprocessing. Data
+exploration and visualization examples are included. Chapters 3 and 4 cover
 classification methods, model selection, model evaluation, different types of
 classifiers, and essential practical issues like class imbalance. Chapter 5
 introduces association analysis with a strong emphasis on visualization.
@@ -74,16 +74,16 @@ cluster evaluation, and the effect of outliers.
 
 # Instructional Design
 
-This resource does not replace the Introduction to Data Mining textbook or
-instruction by a teacher, it instead provides supporting material for learning
+This resource does not replace the _Introduction to Data Mining_ textbook or
+instruction by a teacher. It instead provides supporting material for learning
 to implement data mining concepts in R. The learner is expected to have some
 programming experience and basic statistics knowledge. 
 
-The resource can be used for self-study by any interested person using it to
-accompany reading the Introduction to Data Mining textbook but its main purpose
+The resource can be used for self-study by any interested person together
+with reading the _Introduction to Data Mining_ textbook, but its main purpose
 is to be used as a component for designing an introductory data mining course
 for advanced undergraduate or graduate students. To support instructors, in
-addition to the documented code examples, also complete presentation slide sets
+addition to the documented code examples, complete presentation slide sets
 are provided on the book's GitHub page in PDF and PowerPoint format.  The
 slides are organized in the same way as the resource. A direct connection
 between the slides and the code examples is provided by the R symbol on the
@@ -92,8 +92,8 @@ be studied by the students outside of class or used by the instructor in class.
 
[…]
