### 2024-12-26 #223 @labarba
Congratulations @mhahsler – your JOSE paper is published! 🚀

Huge thanks to our Editor: @stats-tgeorge and Reviewer: @rudeboybert — you make this possible and we are grateful 🙏
